Generates queries that work across registers

From high-level search terms provided by the user, generate specific queries for each registers with which ctrdata works, see ctrdata-registers. Search terms that are expanded to concepts such as from MeSH and MedDRA by the search implementations in registers include the intervention and condition, see ctrdata-registers.

Usage

ctrGenerateQueries(
  condition = NULL,
  intervention = NULL,
  phase = NULL,
  population = NULL,
  recruitment = NULL,
  startBefore = NULL,
  startAfter = NULL,
  completedBefore = NULL,
  completedAfter = NULL,
  onlyWithResults = FALSE,
  registers = c("EUCTR", "ISRCTN", "CTIS", "CTGOV2")
)

Arguments

condition

String with condition / disease

intervention

String with intervention

phase

String, e.g. "phase 2" (note that "phase 2+3" is a specific category, not the union set of "phase 2" and "phase 3")

population

String, e.g. "P" (paediatric), "A" (adult), "P+A" (adult and paediatric), "E" (elderly), "P+A+E" participants can be recruited

recruitment

String, one of "ongoing", "completed", "other"

startBefore

String that can be interpreted as date, see example

startAfter

String that can be interpreted as date

completedBefore

String that can be interpreted as date (does not work with EUCTR)

completedAfter

String that can be interpreted as date (does not work with EUCTR)

onlyWithResults

Logical

registers

Vector of register names, default all registers

Value

Named vector of URLs for finding trials in the registers and as input to functions ctrLoadQueryIntoDb and ctrOpenSearchPagesInBrowser

Examples

urls <- ctrGenerateQueries(
  intervention = "antibody",
  recruitment = "ongoing",
  phase = "phase 3",
  startAfter = "2000-01-01")

# open queries in register web interface
sapply(urls, ctrOpenSearchPagesInBrowser)
#> * Found search query from EUCTR: query=antibody&phase=phase-three&status=ongoing&status=trial-now-transitioned&status=suspended-by-ca&status=temporarily-halted&status=restarted&dateFrom=2000-01-01
#> * Appears specific for CTGOV REST API 2.0
#> * Found search query from CTGOV2: intr=antibody&start=2000-01-01_&aggFilters=phase:3,status:act rec
#> * Found search query from ISRCTN: &filters=intervention:antibody,phase:Phase III,trialStatus:ongoing,GT+overallStartDate:2000-01-01&q=
#> * Found search query from CTIS: searchCriteria={"containAll":"antibody","trialPhaseCode":[5],"status":[2,3,4,6,7],"eeaStartDateFrom":"2000-01-01"}
#>                                                                                                                                                                                                                              EUCTR 
#> "https://www.clinicaltrialsregister.eu/ctr-search/search?query=antibody&phase=phase-three&status=ongoing&status=trial-now-transitioned&status=suspended-by-ca&status=temporarily-halted&status=restarted&dateFrom=2000-01-01#tabs" 
#>                                                                                                                                                                                                                             CTGOV2 
#>                                                                                                                              "https://clinicaltrials.gov/search?intr=antibody&start=2000-01-01_&aggFilters=phase:3,status:act rec" 
#>                                                                                                                                                                                                                             ISRCTN 
#>                                                                                               "https://www.isrctn.com/search?&filters=intervention:antibody,phase:Phase III,trialStatus:ongoing,GT+overallStartDate:2000-01-01&q=" 
#>                                                                                                                                                                                                                               CTIS 
#>                                                    "https://euclinicaltrials.eu/ctis-public/search#searchCriteria={\"containAll\":\"antibody\",\"trialPhaseCode\":[5],\"status\":[2,3,4,6,7],\"eeaStartDateFrom\":\"2000-01-01\"}" 

# load queries into database collection
# sapply(urls, ctrLoadQueryIntoDb, con = dbc)