ctrLoadQueryIntoDbCtgov

ctrLoadQueryIntoDbCtgov

ctrLoadQueryIntoDbCtgov(
  queryterm = queryterm,
  register,
  euctrresults,
  euctrresultshistory,
  annotation.text,
  annotation.mode,
  parallelretrievals,
  only.count,
  con,
  verbose,
  queryupdateterm
)

Arguments

queryterm
Either a string with the full URL of a search in a register or the data frame returned by the ctrGetQueryUrlFromBrowser or the dbQueryHistory functions. The queryterm is recorded in the collection for later use to update records.

register
Vector of abbreviations of the register to query, defaults to "EUCTR".

euctrresults
If TRUE, also download available results when retrieving and loading trials from EUCTR. This slows down this function. (For CTGOV, all available results are always retrieved and loaded.)

euctrresultshistory
If TRUE, also download available history of results publication in EUCTR. This is quite time-consuming (default is FALSE).

annotation.text
Text to be including in the records retrieved with the current query, in the field "annotation".

annotation.mode
One of "append" (default), "prepend" or "replace" for new annotation.text with respect to any existing annotation for the records retreived with the current query.

parallelretrievals
Number of parallel downloads of information from the register, defaults to 10.

only.count
Set to TRUE to return only the number of trial records found in the register for the query. Does not load trial information into the database. Default is FALSE.

con
A src connection object, as obtained with nodbi::src_mongo() or nodbi::src_sqlite()

verbose
Printing additional information if set to TRUE; default is FALSE.