Show structure and all data of a single trial

If used interactively, the function shows a widget of all data in the trial as a tree of field names and values. The widget opens in the default browser. Fields names and values can be search and selected. Selected fields can be copied to the clipboard for use with function dbGetFieldsIntoDf.

Usage

ctrShowOneTrial(identifier = NULL, con = NULL)

Arguments

identifier

A trial identifier string

con

A database connection object, created with nodbi. See section `1 - Database connection` in ctrdata.

Value

Invisibly, the data frame resulting from dfTrials2Long for the trial with the identifier.

Details

If used non-interactively, the data frame is shown that results from dfTrials2Long for the trial with the identifier.

The trial is searched online if not database con is provided or if the trial is not found in the database.

Examples

dbc <- nodbi::src_sqlite(
   dbname = system.file("extdata", "demo.sqlite", package = "ctrdata"),
   collection = "my_trials",
   RSQLite::SQLITE_RO)
#> RSQLite version has enabled accelerating docdb_create() and docdb_update() functions when used with value = <NDJSON file name>.

nodbi::docdb_query(dbc, dbc$collection, "{}", fields = '{"_id":1}')
#>                  _id
#> 1           12949496
#> 2           13281214
#> 3           17473621
#> 4  2012-003632-23-CZ
#> 5  2012-003632-23-SE
#> 6  2014-002606-20-PT
#> 7  2014-003556-31-GB
#> 8  2014-003556-31-SE
#> 9  2022-501142-30-00
#> 10 2023-505613-24-00
#> 11 2024-510663-34-00
#> 12          20343063
#> 13          61070850
#> 14          76463425
#> 15          80181452
#> 16          88261002
#> 17       NCT00617929
#> 18       NCT01125800
#> 19       NCT01483820
#> 20       NCT01505608
#> 21       NCT01592045
#> 22       NCT02620761
#> 23       NCT03280147
#> 24       NCT03325439
#> 25       NCT03431558
#> 26       NCT03876704
#> 27       NCT04001712
#> 28       NCT04041765
#> 29       NCT05969327
#> 30         meta-info

id <- "NCT00617929"
id <- "2012-003632-23"
id <- "80181452"
id <- "2022-501142-30-00"

ctrGetQueryUrl(url = id)
#> * Found search query from CTIS: searchCriteria={"number":"2022-501142-30-00"}
#>                                      query-term query-register
#> 1 searchCriteria={"number":"2022-501142-30-00"}           CTIS

ctrShowOneTrial(identifier = id)
#> * Found search query from CTIS: searchCriteria={"number":"2022-501142-30-00"}
#> Error in utils::browseURL(tf): 'browser' must be a non-empty character string
ctrShowOneTrial(identifier = id, con = dbc)
#> Error in utils::browseURL(tf): 'browser' must be a non-empty character string