Information on clinical trial registers

Registers of the four clinical trial registers from which package ctrdata can retrieve, aggregate and analyse protocol- and result-related information as well as documents, last updated 2026-10-04.

1 - Overview

EUCTR: The EU Clinical Trials Register is complete with more than 44,400 clinical trials (at least one investigational medicinal product, IMP; in the European Union and beyond), including more than 26,200 trials with results, which continue to be added.

CTIS: The EU Clinical Trials Information System, launched in 2023, holds more than 12,500 publicly accessible clinical trials, including more than 1,550 with results or a report. (To automatically get CTIS search query URLs, see here)

CTGOV2: ClinicalTrials.gov holds more than 605,000 interventional and observational studies, including more than 37,400 applicable clinical trials (ACT, e.g. interventional studies with an Investigational Drug or Device, see here ), of which more than 14,400 with results.

ISRCTN: The ISRCTN Registry holds more than 28,400 interventional and observational health studies.

EUCTR	CTIS	CTGOV2	ISRCTN

Protocol-related information	Structured	Structured	Structured	Structured

Result-related information	Structured,
documents	Documents	Structured,
publication links	Publication links

Documents	Reports	Protocol, info sheets,
SAP, results, other	Primarily
protocol, SAP	Protocol, info sheets,
SAP, results, other

2 - Notable changes

EUCTR removed search parameter status= as of February 2025 (see transitioned trials in dbFindIdsUniqueTrials).

CTIS can be used with ctrdata since 2023-03-25. CTIS was relaunched on 2024-06-17, changing the data structure and search syntax, to which ctrdata was updated.

CTGOV "classic" was retired on 2024-06-25; ctrdata subsequently translates CTGOV queries to CTGOV2 queries. The new website ("CTGOV2") can be used with ctrdata since 2023-08-27. Database collections created with CTGOV queries can still be used since functions in ctrdata continue to support them.

More information on changes: here.

3 - References

Material	EUCTR	CTIS	CTGOV2	ISRCTN

About	link	link	link	link

Terms & conditions, disclaimer	link	link	link	link

How to search	link	link	link	link

Search interface	link	link	link	link

Expert / advanced search	link	link	link	link

Glossary / related information	link	link	link	link

FAQ / caveats / examples	link	link	link, link, link	link

Data dictionaries / structure	link, link, link	XLSX files	link, link, link	link

Some registers are expanding entered search terms using dictionaries (example).

4 - Example and ctrdata motivation

The example is an expert search for interventional clinical trials primarily with neonates, investigating medicines for infectious conditions. It shows that searches in the web interface of most registers are not sufficient to identify and analyse the trials of interest:

EUCTR has a search box to retrieve trials with neonates, but not only those conducted exclusively in neonates (link).

CTIS retrieves trials based on their information mentioning the words neonates and infection (link; to show CTIS search results, see here).

CTGOV2 can retrieve trials with neonates only if a specific age definition is used (link).

ISRCTN retrieves studies only with neonates, but not only of medicines (link).

To address these issues, trials can first be retrieved using the links above with ctrLoadQueryIntoDb. In a second step, all trial or a subset can be analysed.

ctrdata supports users with pre-defined ctrdata-trial-concepts that can be calculated across registers when creating a dataframe for analysis, such as f.isMedIntervTrial which identifies the subset of investigational medicine trials.

In addition, all fields made public by the registers can be searched with dbFindFields; the trial's structure and the fields' values can be reviewed with ctrShowOneTrial.

For the example, the following fields can be included when creating a dataframe using dbGetFieldsIntoDf, to select or analyse trials:

EUCTR: f113_newborns_027_days, f1131_number_of_subjects_for_this_age_range

CTIS: authorizedApplication.authorizedPartI.medicalConditions.medicalCondition

CTGOV2: protocolSection.eligibilityModule.maximumAge

ISRCTN: participants.upperAgeLimit

See vignette ctrdata_summarise for several other examples.

Author

Ralf Herold ralf.herold@mailbox.org