Calculate fields from data in other fields

Across registers, calculate a trial concept (e.g., status of recruitment) from fields in a data frame. The concept is defined within ctrdata based on common understanding and any relevant publications. Details of the function are printed to explain it, if no data frame is specified.

Usage

dfCalculate(name = ".*", df = NULL)

Arguments

name

String with name of function to be applied to `df`, or regular expression to list available functions.

df

Optional. Data frame with fields needed to apply function `name`.

Value

Data frame with additional column of name `name` as calculated applying the function `name`. If `df` is `NULL` (default), either a list of functions corresponding to `name` or, if exactly one function is identified, prints details of the function `name` to explain the implementation of the concept and to return invisibly the names of fields needed for its calculation.

Details

The names of the functions can also be used for the argument `calculate` in dbGetFieldsIntoDf to achieve the same purpose already at the time that a data frame with data is generated from the trial collection. This approach will in most cases be more efficient and preferred over using dfCalculate.

Examples

# list names of all available functions in ctrdata
dfCalculate()
#> [1] ".isMedIntervTrial"      ".isPlatformTrial"       ".isUniqueTrial"        
#> [4] ".numTestArmsSubstances" ".startDate"             ".statusRecruitment"    

# list names of functions for regular expression of name
dfCalculate(name = "status")
#> [1] ".statusRecruitment"

# describe a specific function
dfCalculate(name = ".statusRecruitment")
#> 
#> >>>> .statusRecruitment
#> 
#> * Description:
#> 
#> Calculates the status at the time of loading the trial records. Maps the categories that are in fields which specify the state of recruitment. Simplifies them into categories "ongoing", "completed" and "other." Returns an ordered factor.
#> 
#> * Fields needed:
#> 
#> p_end_of_trial_status
#> last_known_status
#> overall_status
#> protocolSection.statusModule.overallStatus
#> participants.recruitmentEnd
#> participants.recruitmentStart
#> participants.recruitmentStatusOverride
#> ctPublicStatus
#> 
#> * To show the implementation, call: ctrdata::.statusRecruitment
#> or, after package("ctrdata"), call: .statusRecruitment
#> 
#> <<<<

# print descriptions of all functions
invisible(sapply(dfCalculate(), dfCalculate))
#> 
#> >>>> .isMedIntervTrial
#> 
#> * Description:
#> 
#> Calculates if record is a medicine-interventional trial, investigating one or more medicine, whether biological or not.
#> 
#> For EUCTR and CTIS, this corresponds to all records as per the definition of the EU Clinical Trial Regulation.
#> 
#> For CTGOV and CTGOV2, this is based on drug or biological as type of intervention, and interventional as type of study.
#> 
#> For ISRCTN, this is based on drug or biological as type of intervention, and interventional as type of study.
#> 
#> Returns a logical.
#> 
#> * Fields needed:
#> 
#> ctrname
#> intervention.intervention_type
#> study_type
#> protocolSection.armsInterventionsModule.interventions.type
#> protocolSection.designModule.studyType
#> interventions.intervention.interventionType
#> trialDesign.primaryStudyDesign
#> ctrname
#> 
#> * To show the implementation, call: ctrdata::.isMedIntervTrial
#> or, after package("ctrdata"), call: .isMedIntervTrial
#> 
#> <<<<
#> 
#> >>>> .isPlatformTrial
#> 
#> * Description:
#> 
#> Calculates if the trial is likely a platform trial or not. As operational definition, at least one of these criteria is true:
#> 
#> - trial has "platform", "basket" or "umbrella" in its title or description (for ISRCTN, this is the only criterion; some trials in EUCTR lack data in English) - trial has more than 2 active arms with different investigational medicines, after excluding comparator, auxiliary and placebo medicines (calculated with function .numTestArmsSubstances()) - trial more than 2 periods, after excluding screening, extension and follow-up periods (for CTGOV and CTGOV2, this criterion requires results-related data)
#> 
#> Requires packages dplyr and stringdist to be installed; stringdist is used for evaluating names of active substances, which are considered similar when the similarity is 0.8 or higher.
#> 
#> Returns a logical.
#> 
#> * Fields needed:
#> 
#> ctrname
#> a3_full_title_of_the_trial
#> trialInformation.fullTitle
#> subjectDisposition.postAssignmentPeriods.postAssignmentPeriod.title
#> subjectDisposition.postAssignmentPeriods.postAssignmentPeriod.arms.arm
#> official_title
#> detailed_description.textblock
#> clinical_results.participant_flow.period_list.period.title
#> arm_group
#> protocolSection.identificationModule.officialTitle
#> protocolSection.descriptionModule.detailedDescription
#> resultsSection.participantFlowModule.periods.title
#> protocolSection.armsInterventionsModule.armGroups
#> trialDescription.title
#> trialDescription.scientificTitle
#> interventions.intervention.interventionType
#> trialDescription.title
#> authorizedApplication.authorizedPartI.trialDetails.clinicalTrialIdentifiers.fullTitle
#> authorizedApplication.authorizedPartI.trialDetails.clinicalTrialIdentifiers.publicTitle
#> authorizedApplication.authorizedPartI.trialDetails.protocolInformation.studyDesign.periodDetails.title
#> authorizedApplication.authorizedPartI.productRoleGroupInfos
#> 
#> * To show the implementation, call: ctrdata::.isPlatformTrial
#> or, after package("ctrdata"), call: .isPlatformTrial
#> 
#> <<<<
#> 
#> >>>> .isUniqueTrial
#> 
#> * Description:
#> 
#> Applies function dbFindIdsUniqueTrials() with its defaults. Returns a logical.
#> 
#> * Fields needed:
#> 
#> _id
#> ctrname
#> a2_eudract_number
#> a52_us_nct_clinicaltrialsgov_registry_number
#> trialInformation.usctnIdentifier
#> a51_isrctn_international_standard_randomised_controlled_trial_number
#> trialInformation.isrctnIdentifier
#> a41_sponsors_protocol_code_number
#> id_info.secondary_id
#> id_info.org_study_id
#> id_info.nct_id
#> id_info.nct_alias
#> externalRefs.eudraCTNumber
#> externalRefs.clinicalTrialsGovNumber
#> isrctn
#> externalRefs.protocolSerialNumber
#> ctNumber
#> eudraCtInfo.eudraCtCode
#> authorizedPartI.trialDetails.clinicalTrialIdentifiers.secondaryIdentifyingNumbers.nctNumber.number
#> protocolSection.identificationModule.nctId
#> protocolSection.identificationModule.secondaryIdInfos.id
#> protocolSection.identificationModule.nctIdAliases
#> protocolSection.identificationModule.orgStudyIdInfo.id
#> 
#> * To show the implementation, call: ctrdata::.isUniqueTrial
#> or, after package("ctrdata"), call: .isUniqueTrial
#> 
#> <<<<
#> 
#> >>>> .numTestArmsSubstances
#> 
#> * Description:
#> 
#> Calculates the number of active arms with different investigational medicines, after excluding comparator, auxiliary and placebo medicines. (No data found for ISRCTN to calculate.)
#> 
#> Requires packages dplyr and stringdist to be installed; stringdist is used for evaluating names of active substances, which are considered similar when the similarity is 0.8 or higher.
#> 
#> Returns an integer.
#> 
#> * Fields needed:
#> 
#> subjectDisposition.postAssignmentPeriods.postAssignmentPeriod.arms.arm
#> arm_group
#> protocolSection.armsInterventionsModule.armGroups
#> trialDescription.title
#> authorizedApplication.authorizedPartI.productRoleGroupInfos
#> 
#> * To show the implementation, call: ctrdata::.numTestArmsSubstances
#> or, after package("ctrdata"), call: .numTestArmsSubstances
#> 
#> <<<<
#> 
#> >>>> .startDate
#> 
#> * Description:
#> 
#> Calculates the date of start of the trial, based on the documented or planned start of recruitment, or on the date of opinion of the competent authority. Returns a date.
#> 
#> * Fields needed:
#> 
#> n_date_of_competent_authority_decision
#> trialInformation.recruitmentStartDate
#> start_date
#> protocolSection.statusModule.startDateStruct.date
#> participants.recruitmentStart
#> trialDesign.overallStartDate
#> startDateEU
#> authorizedApplication.authorizedPartI.trialDetails.trialInformation.trialDuration.estimatedRecruitmentStartDate
#> 
#> * To show the implementation, call: ctrdata::.startDate
#> or, after package("ctrdata"), call: .startDate
#> 
#> <<<<
#> 
#> >>>> .statusRecruitment
#> 
#> * Description:
#> 
#> Calculates the status at the time of loading the trial records. Maps the categories that are in fields which specify the state of recruitment. Simplifies them into categories "ongoing", "completed" and "other." Returns an ordered factor.
#> 
#> * Fields needed:
#> 
#> p_end_of_trial_status
#> last_known_status
#> overall_status
#> protocolSection.statusModule.overallStatus
#> participants.recruitmentEnd
#> participants.recruitmentStart
#> participants.recruitmentStatusOverride
#> ctPublicStatus
#> 
#> * To show the implementation, call: ctrdata::.statusRecruitment
#> or, after package("ctrdata"), call: .statusRecruitment
#> 
#> <<<<

# apply dfCalculate to data frame
dbc <- nodbi::src_sqlite(
   dbname = system.file("extdata", "demo.sqlite", package = "ctrdata"),
   collection = "my_trials",
   RSQLite::SQLITE_RO)
#> RSQLite version has enabled accelerating docdb_create() and docdb_update() functions when used with value = <NDJSON file name>.

# use with existing data frame,
# first get fields needed
trialsDf <- dbGetFieldsIntoDf(
  fields = unlist(dfCalculate(name = ".statusRecruitment")),
  con = dbc)
#> 
#> >>>> .statusRecruitment
#> 
#> * Description:
#> 
#> Calculates the status at the time of loading the trial records. Maps the categories that are in fields which specify the state of recruitment. Simplifies them into categories "ongoing", "completed" and "other." Returns an ordered factor.
#> 
#> * Fields needed:
#> 
#> p_end_of_trial_status
#> last_known_status
#> overall_status
#> protocolSection.statusModule.overallStatus
#> participants.recruitmentEnd
#> participants.recruitmentStart
#> participants.recruitmentStatusOverride
#> ctPublicStatus
#> 
#> * To show the implementation, call: ctrdata::.statusRecruitment
#> or, after package("ctrdata"), call: .statusRecruitment
#> 
#> <<<<

# then calculate
dfCalculate(
  name = ".statusRecruitment",
  df = trialsDf)
#>                  _id p_end_of_trial_status  last_known_status overall_status
#> 1           12949496                  <NA>               <NA>           <NA>
#> 2           13281214                  <NA>               <NA>           <NA>
#> 3           17473621                  <NA>               <NA>           <NA>
#> 4  2012-003632-23-CZ             Completed               <NA>           <NA>
#> 5  2012-003632-23-SE             Completed               <NA>           <NA>
#> 6  2014-002606-20-PT               Ongoing               <NA>           <NA>
#> 7  2014-003556-31-SE             Completed               <NA>           <NA>
#> 8  2022-501142-30-00                  <NA>               <NA>           <NA>
#> 9  2023-505613-24-00                  <NA>               <NA>           <NA>
#> 10 2024-510663-34-00                  <NA>               <NA>           <NA>
#> 11          20343063                  <NA>               <NA>           <NA>
#> 12          61070850                  <NA>               <NA>           <NA>
#> 13          76463425                  <NA>               <NA>           <NA>
#> 14          80181452                  <NA>               <NA>           <NA>
#> 15          88261002                  <NA>               <NA>           <NA>
#> 16       NCT00617929                  <NA>               <NA>           <NA>
#> 17       NCT01125800                  <NA>               <NA>           <NA>
#> 18       NCT01483820                  <NA>               <NA>           <NA>
#> 19       NCT01505608                  <NA>               <NA>           <NA>
#> 20       NCT01592045                  <NA>               <NA>           <NA>
#> 21       NCT02620761                  <NA>               <NA>     Terminated
#> 22       NCT03280147                  <NA>               <NA>      Completed
#> 23       NCT03325439                  <NA>               <NA>     Terminated
#> 24       NCT03431558                  <NA>         Recruiting Unknown status
#> 25       NCT03876704                  <NA>         Recruiting Unknown status
#> 26       NCT04001712                  <NA>               <NA>      Completed
#> 27       NCT04041765                  <NA> Not yet recruiting Unknown status
#> 28       NCT05969327                  <NA>               <NA>      Completed
#>    protocolSection.statusModule.overallStatus participants.recruitmentEnd
#> 1                                        <NA>                  2020-01-01
#> 2                                        <NA>                  2020-01-01
#> 3                                        <NA>                  2024-02-01
#> 4                                        <NA>                        <NA>
#> 5                                        <NA>                        <NA>
#> 6                                        <NA>                        <NA>
#> 7                                        <NA>                        <NA>
#> 8                                        <NA>                        <NA>
#> 9                                        <NA>                        <NA>
#> 10                                       <NA>                        <NA>
#> 11                                       <NA>                  2013-12-31
#> 12                                       <NA>                  2011-12-31
#> 13                                       <NA>                  2015-05-31
#> 14                                       <NA>                  2012-03-31
#> 15                                       <NA>                  2017-09-28
#> 16                                 TERMINATED                        <NA>
#> 17                                  COMPLETED                        <NA>
#> 18                                 TERMINATED                        <NA>
#> 19                                 TERMINATED                        <NA>
#> 20                                  COMPLETED                        <NA>
#> 21                                       <NA>                        <NA>
#> 22                                       <NA>                        <NA>
#> 23                                       <NA>                        <NA>
#> 24                                       <NA>                        <NA>
#> 25                                       <NA>                        <NA>
#> 26                                       <NA>                        <NA>
#> 27                                       <NA>                        <NA>
#> 28                                       <NA>                        <NA>
#>    participants.recruitmentStart participants.recruitmentStatusOverride
#> 1                     2015-11-01                                Stopped
#> 2                     2016-09-01                                   <NA>
#> 3                     2023-05-01                                   <NA>
#> 4                           <NA>                                   <NA>
#> 5                           <NA>                                   <NA>
#> 6                           <NA>                                   <NA>
#> 7                           <NA>                                   <NA>
#> 8                           <NA>                                   <NA>
#> 9                           <NA>                                   <NA>
#> 10                          <NA>                                   <NA>
#> 11                    2011-01-01                                   <NA>
#> 12                    2007-01-10                                   <NA>
#> 13                    2013-06-01                                   <NA>
#> 14                    2010-04-01                                   <NA>
#> 15                    2013-09-01                                   <NA>
#> 16                          <NA>                                   <NA>
#> 17                          <NA>                                   <NA>
#> 18                          <NA>                                   <NA>
#> 19                          <NA>                                   <NA>
#> 20                          <NA>                                   <NA>
#> 21                          <NA>                                   <NA>
#> 22                          <NA>                                   <NA>
#> 23                          <NA>                                   <NA>
#> 24                          <NA>                                   <NA>
#> 25                          <NA>                                   <NA>
#> 26                          <NA>                                   <NA>
#> 27                          <NA>                                   <NA>
#> 28                          <NA>                                   <NA>
#>         ctPublicStatus .statusRecruitment
#> 1                 <NA>              other
#> 2                 <NA>          completed
#> 3                 <NA>          completed
#> 4                 <NA>          completed
#> 5                 <NA>          completed
#> 6                 <NA>            ongoing
#> 7                 <NA>          completed
#> 8  Ongoing, recruiting            ongoing
#> 9  Ongoing, recruiting            ongoing
#> 10 Ongoing, recruiting            ongoing
#> 11                <NA>          completed
#> 12                <NA>          completed
#> 13                <NA>          completed
#> 14                <NA>          completed
#> 15                <NA>          completed
#> 16                <NA>              other
#> 17                <NA>          completed
#> 18                <NA>              other
#> 19                <NA>              other
#> 20                <NA>          completed
#> 21                <NA>              other
#> 22                <NA>          completed
#> 23                <NA>              other
#> 24                <NA>            ongoing
#> 25                <NA>            ongoing
#> 26                <NA>          completed
#> 27                <NA>            ongoing
#> 28                <NA>          completed

# or use already when creating a trial data frame
# to subset to the unique records of clinical trials
trialsDf <- dbGetFieldsIntoDf(
  calculate = ".isUniqueTrial",
  con = dbc)
#> - Finding duplicates among registers' and sponsor ids...
#> - 2 EUCTR _id were not preferred EU Member State record for 8 trials
#> - Keeping 3 / 8 / 5 / 8 / 3 records from EUCTR / CTGOV / CTGOV2 / ISRCTN / CTIS
trialsDf[trialsDf[[".isUniqueTrial"]], ]
#>                  _id .isUniqueTrial
#> 1           12949496           TRUE
#> 2           13281214           TRUE
#> 3           17473621           TRUE
#> 5  2012-003632-23-SE           TRUE
#> 6  2014-002606-20-PT           TRUE
#> 8  2014-003556-31-SE           TRUE
#> 9  2022-501142-30-00           TRUE
#> 10 2023-505613-24-00           TRUE
#> 11 2024-510663-34-00           TRUE
#> 12          20343063           TRUE
#> 13          61070850           TRUE
#> 14          76463425           TRUE
#> 15          80181452           TRUE
#> 16          88261002           TRUE
#> 17       NCT00617929           TRUE
#> 18       NCT01125800           TRUE
#> 19       NCT01483820           TRUE
#> 20       NCT01505608           TRUE
#> 21       NCT01592045           TRUE
#> 22       NCT02620761           TRUE
#> 23       NCT03280147           TRUE
#> 24       NCT03325439           TRUE
#> 25       NCT03431558           TRUE
#> 26       NCT03876704           TRUE
#> 27       NCT04001712           TRUE
#> 28       NCT04041765           TRUE
#> 29       NCT05969327           TRUE