Calculate trial concept from various fields and registers

Across registers, calculate a trial concept (e.g., status of recruitment). The concept is defined within ctrdata based on common understanding and any relevant publications. Details of the implementation are printed, if no data frame is specified, to show the peculiarities of the registers for calculating the concept.

Usage

dfCalculateConcept(name = ".*", df = NULL)

Arguments

name

String with name of a function to be applied to `df`, or regular expression to list available functions.

df

Optional. Data frame with fields needed to apply function `name`.

Value

Data frame with additional column of name `name` as calculated applying the function `name`. If `df` is `NULL` (default), either a list of functions corresponding to `name` or, if exactly one function is identified, prints details of the function `name` to explain the implementation of the concept and to return invisibly the names of fields needed for its calculation.

Details

The names of the functions can also be used for the argument `calculate` in dbGetFieldsIntoDf to achieve the same purpose already at the time that a data frame with data is generated from the trial collection. This approach will in most cases be more efficient and preferred over using dfCalculateConcept.

Note that the functions do not operate on historic versions (CTGOV2) but only on information from the latest record.

Examples

# list names of all available functions in ctrdata
dfCalculateConcept()
#>  [1] ".controlType"                ".isMedIntervTrial"          
#>  [3] ".isPlatformTrial"            ".isUniqueTrial"             
#>  [5] ".numSites"                   ".numTestArmsSubstances"     
#>  [7] ".primaryEndpointDescription" ".primaryEndpointResults"    
#>  [9] ".resultsDate"                ".sampleSize"                
#> [11] ".sponsorType"                ".startDate"                 
#> [13] ".statusRecruitment"          ".trialObjectives"           
#> [15] ".trialPhase"                 ".trialPopulation"           

# list names of functions for regular expression of name
dfCalculateConcept(name = "status")
#> [1] ".statusRecruitment"

# describe a specific function
dfCalculateConcept(name = ".statusRecruitment")
#> 
#> >>>> .statusRecruitment
#> 
#> * Description:
#> 
#> Calculates the status at the time of loading the trial records. Maps the categories that are in fields which specify the state of recruitment. Simplifies the status into three categories "ongoing" (includes active, not yet recruiting; temporarily halted; suspended; authorised, not started and similar), "completed" (includes ended; ongoing, recuitment mended) and "other" (includes terminated; revoked).
#> 
#> Returns a factor of these categories.
#> 
#> * Fields needed:
#> 
#> trialInformation.globalEndOfTrialPremature
#> trialInformation.isGlobalEndOfTrialReached
#> p_end_of_trial_status
#> last_known_status
#> overall_status
#> protocolSection.statusModule.overallStatus
#> participants.recruitmentEnd
#> participants.recruitmentStart
#> participants.recruitmentStatusOverride
#> ctPublicStatusCode
#> ctStatus
#> 
#> * To show the implementation, call: ctrdata::.statusRecruitment
#> or, after package("ctrdata"), call: .statusRecruitment
#> 
#> <<<<

# print descriptions of all functions
invisible(sapply(dfCalculateConcept(), dfCalculateConcept))
#> 
#> >>>> .controlType
#> 
#> * Description:
#> 
#> Calculates the type of control(s) data collected within a clinical trial. ICH E10 lists as types of control: placebo concurrent control, no-treatment concurrent control, dose-response concurrent control, active (positive) concurrent control, external (including historical) control, multiple control groups.
#> 
#> Dose-controlled trials are currently not identified. External (including historical) controls are so far not identified in specific register fields. Cross-over designs, where identifiable, have active controls.
#> 
#> Returns a factor with levels "none", "no-treatment", "placebo", "active", "placebo+active" and "other".
#> 
#> * Fields needed:
#> 
#> e81_controlled
#> e816_cross_over
#> e817_other
#> e8171_other_trial_design_description
#> e822_placebo
#> e823_other
#> e8231_comparator_description
#> e824_number_of_treatment_arms_in_the_trial
#> arm_group.arm_group_type
#> protocolSection.designModule.designInfo.interventionModel
#> protocolSection.armsInterventionsModule.armGroups.type
#> trialDesign.studyDesign
#> trialDesign.primaryStudyDesign
#> authorizedPartI.productRoleGroupInfos.productRoleName
#> authorizedApplication.authorizedPartI.productRoleGroupInfos.productRoleName
#> 
#> * To show the implementation, call: ctrdata::.controlType
#> or, after package("ctrdata"), call: .controlType
#> 
#> <<<<
#> 
#> >>>> .isMedIntervTrial
#> 
#> * Description:
#> 
#> Calculates if record is a medicine-interventional trial, investigating one or more medicine, whether biological or not.
#> 
#> For EUCTR and CTIS, this corresponds to all records as per the definition of the EU Clinical Trial Regulation.
#> 
#> For CTGOV and CTGOV2, this is based on drug or biological as type of intervention, and interventional as type of study.
#> 
#> For ISRCTN, this is based on drug or biological as type of intervention, and interventional as type of study.
#> 
#> Returns a logical.
#> 
#> * Fields needed:
#> 
#> ctrname
#> intervention.intervention_type
#> study_type
#> protocolSection.armsInterventionsModule.interventions.type
#> protocolSection.designModule.studyType
#> interventions.intervention.interventionType
#> trialDesign.primaryStudyDesign
#> ctrname
#> 
#> * To show the implementation, call: ctrdata::.isMedIntervTrial
#> or, after package("ctrdata"), call: .isMedIntervTrial
#> 
#> <<<<
#> 
#> >>>> .isPlatformTrial
#> 
#> * Description:
#> 
#> Calculates if the trial is likely a platform trial or not. As operational definition, at least one of these criteria is true:
#> 
#> - trial has "platform", "basket", "umbrella", "multi-?arm" or "multi-?stage" in its title or description (for ISRCTN, this is the only criterion; some trials in EUCTR lack data in English) - trial has more than 2 active arms with different investigational medicines, after excluding comparator, auxiliary and placebo medicines (calculated with function .numTestArmsSubstances(), not used for ISRCTN because it cannot be calculated precisely) - trial more than 2 periods, after excluding safety run-in, screening, enrolling, extension and follow-up periods (for CTGOV and CTGOV2, this criterion requires results-related data)
#> 
#> Requires that EUCTR results have been included in the collection, using ctrLoadQueryIntoDb(queryterm = ..., euctrresults = TRUE, con = ...).
#> 
#> Requires packages dplyr and stringdist to be installed; stringdist is used for evaluating names of active substances, which are considered similar when the similarity is 0.8 or higher.
#> 
#> Returns a logical.
#> 
#> * Fields needed:
#> 
#> ctrname
#> a3_full_title_of_the_trial
#> trialInformation.fullTitle
#> subjectDisposition.postAssignmentPeriods.postAssignmentPeriod.title
#> e824_number_of_treatment_arms_in_the_trial
#> e81_controlled
#> subjectDisposition.postAssignmentPeriods.postAssignmentPeriod.arms.arm
#> official_title
#> detailed_description.textblock
#> clinical_results.participant_flow.period_list.period.title
#> arm_group
#> protocolSection.identificationModule.officialTitle
#> protocolSection.descriptionModule.detailedDescription
#> resultsSection.participantFlowModule.periods.title
#> protocolSection.armsInterventionsModule.armGroups
#> trialDescription.title
#> trialDescription.scientificTitle
#> interventions.intervention.interventionType
#> trialDesign.secondaryStudyDesign
#> interventions.intervention.interventionType
#> interventions.intervention.drugNames
#> title
#> applications.fullTitle
#> authorizedPartI.trialDetails.clinicalTrialIdentifiers.fullTitle
#> authorizedPartI.trialDetails.clinicalTrialIdentifiers.publicTitle
#> authorizedPartI.trialDetails.protocolInformation.studyDesign.periodDetails.title
#> authorizedPartI.trialDetails.protocolInformation.studyDesign.periodDetails.armDetails.title
#> authorizedApplication.authorizedPartI.trialDetails.clinicalTrialIdentifiers.fullTitle
#> authorizedApplication.authorizedPartI.trialDetails.clinicalTrialIdentifiers.publicTitle
#> authorizedApplication.authorizedPartI.trialDetails.protocolInformation.studyDesign.periodDetails.title
#> authorizedPartI.productRoleGroupInfos
#> authorizedApplication.authorizedPartI.productRoleGroupInfos
#> 
#> * To show the implementation, call: ctrdata::.isPlatformTrial
#> or, after package("ctrdata"), call: .isPlatformTrial
#> 
#> <<<<
#> 
#> >>>> .isUniqueTrial
#> 
#> * Description:
#> 
#> Applies function dbFindIdsUniqueTrials() with its defaults. Returns a logical.
#> 
#> * Fields needed:
#> 
#> _id
#> ctrname
#> a2_eudract_number
#> a52_us_nct_clinicaltrialsgov_registry_number
#> trialInformation.usctnIdentifier
#> a51_isrctn_international_standard_randomised_controlled_trial_number
#> trialInformation.isrctnIdentifier
#> a41_sponsors_protocol_code_number
#> id_info.secondary_id
#> id_info.org_study_id
#> id_info.nct_id
#> id_info.nct_alias
#> externalRefs.eudraCTNumber
#> externalRefs.clinicalTrialsGovNumber
#> isrctn
#> externalRefs.protocolSerialNumber
#> ctNumber
#> eudraCtInfo.eudraCtCode
#> authorizedPartI.trialDetails.clinicalTrialIdentifiers.secondaryIdentifyingNumbers.nctNumber.number
#> protocolSection.identificationModule.nctId
#> protocolSection.identificationModule.secondaryIdInfos.id
#> protocolSection.identificationModule.nctIdAliases
#> protocolSection.identificationModule.orgStudyIdInfo.id
#> 
#> * To show the implementation, call: ctrdata::.isUniqueTrial
#> or, after package("ctrdata"), call: .isUniqueTrial
#> 
#> <<<<
#> 
#> >>>> .numSites
#> 
#> * Description:
#> 
#> Calculates the number of the sites where the trial is conducted. EUCTR lacks information on number of sites outside of the EEA; for each non-EEA country mentioned, at least one site is assumed.
#> 
#> Returns an integer.
#> 
#> * Fields needed:
#> 
#> e851_number_of_sites_anticipated_in_the_eea
#> e863_trial_sites_planned_in
#> location.facility.address.city
#> protocolSection.contactsLocationsModule.locations.city
#> participants.trialCentres.trialCentre.name
#> authorizedPartsII.trialSites.id
#> authorizedApplication.authorizedPartsII.trialSites.id
#> 
#> * To show the implementation, call: ctrdata::.numSites
#> or, after package("ctrdata"), call: .numSites
#> 
#> <<<<
#> 
#> >>>> .numTestArmsSubstances
#> 
#> * Description:
#> 
#> Calculates the number of active arms with different investigational medicines, after excluding comparator, auxiliary and placebo arms / medicines. For ISRCTN, this is imprecise because arms are not identified in a field. Most registers provide no or only limited information on phase 1 trials, so that this number typically cannot be calculated for these trials.
#> 
#> Requires packages stringdist to be installed; stringdist is used for evaluating names of active substances, which are considered similar when the similarity is 0.8 or higher.
#> 
#> Returns an integer.
#> 
#> * Fields needed:
#> 
#> e824_number_of_treatment_arms_in_the_trial
#> e81_controlled
#> subjectDisposition.postAssignmentPeriods.postAssignmentPeriod.arms.arm
#> arm_group
#> protocolSection.armsInterventionsModule.armGroups
#> trialDesign.secondaryStudyDesign
#> interventions.intervention.interventionType
#> interventions.intervention.drugNames
#> authorizedPartI.productRoleGroupInfos
#> authorizedApplication.authorizedPartI.productRoleGroupInfos
#> 
#> * To show the implementation, call: ctrdata::.numTestArmsSubstances
#> or, after package("ctrdata"), call: .numTestArmsSubstances
#> 
#> <<<<
#> 
#> >>>> .primaryEndpointDescription
#> 
#> * Description:
#> 
#> Calculates the full description of the primary endpoint, including e.g. its title, description, timeframe of assessment. The details vary by register. The text description can be used for identifying trials of interest or for analysing trends in primary endpoints, which among the set of all endpoints are most often used for determining the number of participants sought for the study.
#> 
#> Returns a list (that is, one or more items in one vector per row; the reason is a proportion of records lists several endpoints as primary).
#> 
#> * Fields needed:
#> 
#> e51_primary_end_points
#> e511_timepoints_of_evaluation_of_this_end_point
#> primary_outcome.measure
#> primary_outcome.description
#> primary_outcome.time_frame
#> protocolSection.outcomesModule.primaryOutcomes.measure
#> protocolSection.outcomesModule.primaryOutcomes.description
#> protocolSection.outcomesModule.primaryOutcomes.timeFrame
#> trialDescription.primaryOutcome
#> authorizedApplication.authorizedPartI.trialDetails.trialInformation.endPoint.primaryEndPoints.endPoint
#> authorizedPartI.trialDetails.trialInformation.endPoint.primaryEndPoints.endPoint
#> 
#> * To show the implementation, call: ctrdata::.primaryEndpointDescription
#> or, after package("ctrdata"), call: .primaryEndpointDescription
#> 
#> <<<<
#> 
#> >>>> .primaryEndpointResults
#> 
#> * Description:
#> 
#> Calculates several results-related elements of the primary analysis of the primary endpoint. Requires loading results-related information. For CTIS and ISRCTN, such information is not available in a structured format.
#> 
#> Recommended to be combined with .controlType, .sampleSize etc. for analyses.
#> 
#> Returns new columns: .primaryEndpointFirstPvalue (discarding any inequality indicator, e.g. <=) .primaryEndpointFirstPmethod (normalised string, e.g. chisquared) .primaryEndpointFirstPsize (number included in test, across assignment groups)
#> 
#> * Fields needed:
#> 
#> endPoints.endPoint.statisticalAnalyses.statisticalAnalysis.statisticalHypothesisTest.value
#> endPoints.endPoint.statisticalAnalyses.statisticalAnalysis.statisticalHypothesisTest.method.value
#> endPoints.endPoint.type.value
#> endPoints.endPoint
#> clinical_results.outcome_list.outcome.analysis_list.analysis.method
#> clinical_results.outcome_list.outcome.analysis_list.analysis.p_value
#> clinical_results.outcome_list.outcome.type
#> clinical_results.outcome_list.outcome
#> resultsSection.outcomeMeasuresModule.outcomeMeasures.analyses.pValue
#> resultsSection.outcomeMeasuresModule.outcomeMeasures.analyses.statisticalMethod
#> resultsSection.outcomeMeasuresModule.outcomeMeasures.type
#> resultsSection.outcomeMeasuresModule.outcomeMeasures
#> 
#> * To show the implementation, call: ctrdata::.primaryEndpointResults
#> or, after package("ctrdata"), call: .primaryEndpointResults
#> 
#> <<<<
#> 
#> >>>> .resultsDate
#> 
#> * Description:
#> 
#> Calculates the earliest date of results as recorded in the register. At that date, results may have been incomplete and may have been changed later. For EUCTR, requires that results and preferrably also their history of publication have been included in the collection, using ctrLoadQueryIntoDb(queryterm = ..., euctrresults{history} = TRUE, con = ...). Cannot be calculated for ISRCTN, which does not have a corresponding field.
#> 
#> Returns a date.
#> 
#> * Fields needed:
#> 
#> firstreceived_results_date
#> trialInformation.analysisStageDate
#> results_first_posted
#> protocolSection.statusModule.resultsFirstPostDateStruct.date
#> results.intentToPublish
#> results.summaryResults.submissionDate
#> results.clinicalStudyReports.submitDate
#> 
#> * To show the implementation, call: ctrdata::.resultsDate
#> or, after package("ctrdata"), call: .resultsDate
#> 
#> <<<<
#> 
#> >>>> .sampleSize
#> 
#> * Description:
#> 
#> Calculates the sample size of the trial, preferring results-related over protocol-related information.
#> 
#> Returns an integer.
#> 
#> * Fields needed:
#> 
#> trialInformation.countrySubjectCounts.countrySubjectCount.subjects
#> f11_number_of_subjects_for_this_age_range
#> f1111_number_of_subjects_for_this_age_range
#> f1121_number_of_subjects_for_this_age_range
#> f1131_number_of_subjects_for_this_age_range
#> f1141_number_of_subjects_for_this_age_range
#> f1151_number_of_subjects_for_this_age_range
#> f1161_number_of_subjects_for_this_age_range
#> f121_number_of_subjects_for_this_age_range
#> f131_number_of_subjects_for_this_age_range
#> enrollment
#> protocolSection.designModule.enrollmentInfo.count
#> participants.targetEnrolment
#> participants.totalFinalEnrolment
#> authorizedPartsII.recruitmentSubjectCount
#> authorizedPartI.rowSubjectCount
#> authorizedApplication.authorizedPartsII.recruitmentSubjectCount
#> authorizedApplication.authorizedPartI.rowSubjectCount
#> 
#> * To show the implementation, call: ctrdata::.sampleSize
#> or, after package("ctrdata"), call: .sampleSize
#> 
#> <<<<
#> 
#> >>>> .sponsorType
#> 
#> * Description:
#> 
#> Calculates the type or class of the lead or main sponsor of the trial.
#> 
#> Some information is not yet mapped (e.g., "NETWORK" in CTGOV2). No specific field is available in ISRCTN (which retrieves information on the type of funder via http://data.crossref.org/fundingdata/registry).
#> 
#> Returns a factor with levels "For profit", "Not for profit", "Other"
#> 
#> * Fields needed:
#> 
#> b1_sponsor.b31_and_b32_status_of_the_sponsor
#> sponsors.lead_sponsor.agency_class
#> protocolSection.sponsorCollaboratorsModule.leadSponsor.class
#> ctrname
#> sponsorType
#> primarySponsor.commercial
#> authorizedApplication.authorizedPartI.sponsors.isCommercial
#> 
#> * To show the implementation, call: ctrdata::.sponsorType
#> or, after package("ctrdata"), call: .sponsorType
#> 
#> <<<<
#> 
#> >>>> .startDate
#> 
#> * Description:
#> 
#> Calculates the date of start of the trial, based on the documented or planned start of recruitment, or on the date of opinion of the competent authority. Returns a date.
#> 
#> * Fields needed:
#> 
#> n_date_of_competent_authority_decision
#> n_date_of_ethics_committee_opinion
#> trialInformation.recruitmentStartDate
#> start_date
#> protocolSection.statusModule.startDateStruct.date
#> participants.recruitmentStart
#> trialDesign.overallStartDate
#> startDateEU
#> authorizationDate
#> authorizedApplication.authorizedPartI.trialDetails.trialInformation.trialDuration.estimatedRecruitmentStartDate
#> 
#> * To show the implementation, call: ctrdata::.startDate
#> or, after package("ctrdata"), call: .startDate
#> 
#> <<<<
#> 
#> >>>> .statusRecruitment
#> 
#> * Description:
#> 
#> Calculates the status at the time of loading the trial records. Maps the categories that are in fields which specify the state of recruitment. Simplifies the status into three categories "ongoing" (includes active, not yet recruiting; temporarily halted; suspended; authorised, not started and similar), "completed" (includes ended; ongoing, recuitment mended) and "other" (includes terminated; revoked).
#> 
#> Returns a factor of these categories.
#> 
#> * Fields needed:
#> 
#> trialInformation.globalEndOfTrialPremature
#> trialInformation.isGlobalEndOfTrialReached
#> p_end_of_trial_status
#> last_known_status
#> overall_status
#> protocolSection.statusModule.overallStatus
#> participants.recruitmentEnd
#> participants.recruitmentStart
#> participants.recruitmentStatusOverride
#> ctPublicStatusCode
#> ctStatus
#> 
#> * To show the implementation, call: ctrdata::.statusRecruitment
#> or, after package("ctrdata"), call: .statusRecruitment
#> 
#> <<<<
#> 
#> >>>> .trialObjectives
#> 
#> * Description:
#> 
#> Calculates objectives of the trial, by searching for text fragments found in fields describing its purpose, objective, background or hypothesis, after applying .isMedIntervTrial, because the text fragments are tailored to medicinal product interventional trials.
#> 
#> This is a simplification, and it is expected that the criteria will be further refined. The text fragments only apply to English.
#> 
#> Returns a string with letters separated by a space, such as E (efficacy, including cure, survival, effectiveness); A (activity, including reponse, remission, seroconversion); S (safety); PK; PD (including biomarker); D (dose-finding, determining recommended dose); LT (long-term); and FU (follow-up).
#> 
#> * Fields needed:
#> 
#> e21_main_objective_of_the_trial
#> ctrname
#> detailed_description.textblock
#> study_design_info.primary_purpose
#> intervention.intervention_type
#> study_type
#> protocolSection.descriptionModule.detailedDescription
#> protocolSection.designModule.designInfo.primaryPurpose
#> protocolSection.armsInterventionsModule.interventions.type
#> protocolSection.designModule.studyType
#> trialDescription.studyHypothesis
#> trialDesign.trialType
#> interventions.intervention.interventionType
#> trialDesign.primaryStudyDesign
#> authorizedPartI.trialDetails.trialInformation.trialObjective.mainObjective
#> authorizedPartI.trialDetails.trialInformation.trialObjective.trialScopes.otherDescription
#> authorizedApplication.authorizedPartI.trialDetails.trialInformation.trialObjective.mainObjective
#> authorizedApplication.authorizedPartI.trialDetails.trialInformation.trialObjective.trialScopes.otherDescription
#> ctrname
#> 
#> * To show the implementation, call: ctrdata::.trialObjectives
#> or, after package("ctrdata"), call: .trialObjectives
#> 
#> <<<<
#> 
#> >>>> .trialPhase
#> 
#> * Description:
#> 
#> Calculates the phase of a clinical trial as per ICH E8(R1).
#> 
#> Returns an ordered factor of levels "phase 1", "phase 1+2", "phase 2", "phase 2+3", "phase 2+4", "phase 3", "phase 3+4", "phase 1+2+3",  "phase 4", "phase 1+2+3+4".
#> 
#> * Fields needed:
#> 
#> e71_human_pharmacology_phase_i
#> e72_therapeutic_exploratory_phase_ii
#> e73_therapeutic_confirmatory_phase_iii
#> e74_therapeutic_use_phase_iv
#> phase
#> protocolSection.designModule.phases
#> interventions.intervention.phase
#> authorizedPartI.trialDetails.trialInformation.trialCategory.trialPhase
#> authorizedApplication.authorizedPartI.trialDetails.trialInformation.trialCategory.trialPhase
#> 
#> * To show the implementation, call: ctrdata::.trialPhase
#> or, after package("ctrdata"), call: .trialPhase
#> 
#> <<<<
#> 
#> >>>> .trialPopulation
#> 
#> * Description:
#> 
#> Calculates from protocol-related information the age groups that can participate in a trial.
#> 
#> Returns new columns: .trialPopulationAgeGroup (factor, "P", "A", "P+A", "E", "A+E", "P+A+E") .trialPopulationInclusion (string) .trialPopulationExclusion (string)
#> 
#> * Fields needed:
#> 
#> e3_principal_inclusion_criteria
#> e4_principal_exclusion_criteria
#> f111_in_utero
#> f112_preterm_newborn_infants_up_to_gestational_age__37_weeks
#> f113_newborns_027_days
#> f114_infants_and_toddlers_28_days23_months
#> f115_children_211years
#> f116_adolescents_1217_years
#> f11_trial_has_subjects_under_18
#> f12_adults_1864_years
#> f13_elderly_65_years
#> eligibility.criteria.textblock
#> eligibility.maximum_age
#> eligibility.minimum_age
#> protocolSection.eligibilityModule.maximumAge
#> protocolSection.eligibilityModule.minimumAge
#> protocolSection.eligibilityModule.eligibilityCriteria
#> participants.ageRange
#> participants.inclusion
#> participants.exclusion
#> ageGroup
#> authorizedApplication.authorizedPartI.trialDetails.trialInformation.eligibilityCriteria.principalInclusionCriteria.principalInclusionCriteria
#> authorizedApplication.authorizedPartI.trialDetails.trialInformation.eligibilityCriteria.principalExclusionCriteria.principalExclusionCriteria
#> 
#> * To show the implementation, call: ctrdata::.trialPopulation
#> or, after package("ctrdata"), call: .trialPopulation
#> 
#> <<<<

# apply dfCalculateConcept to data frame
dbc <- nodbi::src_sqlite(
  dbname = system.file("extdata", "demo.sqlite", package = "ctrdata"),
  collection = "my_trials",
  flags = RSQLite::SQLITE_RO)
#> RSQLite version has enabled accelerating docdb_create() and docdb_update() functions when used with value = <NDJSON file name>.

# use with existing data frame,
# first get fields needed
trialsDf <- dbGetFieldsIntoDf(
  fields = unlist(dfCalculateConcept(name = ".statusRecruitment")),
  con = dbc)
#> 
#> >>>> .statusRecruitment
#> 
#> * Description:
#> 
#> Calculates the status at the time of loading the trial records. Maps the categories that are in fields which specify the state of recruitment. Simplifies the status into three categories "ongoing" (includes active, not yet recruiting; temporarily halted; suspended; authorised, not started and similar), "completed" (includes ended; ongoing, recuitment mended) and "other" (includes terminated; revoked).
#> 
#> Returns a factor of these categories.
#> 
#> * Fields needed:
#> 
#> trialInformation.globalEndOfTrialPremature
#> trialInformation.isGlobalEndOfTrialReached
#> p_end_of_trial_status
#> last_known_status
#> overall_status
#> protocolSection.statusModule.overallStatus
#> participants.recruitmentEnd
#> participants.recruitmentStart
#> participants.recruitmentStatusOverride
#> ctPublicStatusCode
#> ctStatus
#> 
#> * To show the implementation, call: ctrdata::.statusRecruitment
#> or, after package("ctrdata"), call: .statusRecruitment
#> 
#> <<<<
#> Querying database (11 fields)...

# then calculate
dfCalculateConcept(
  name = ".statusRecruitment",
  df = trialsDf)
#> Calculating .statusRecruitment...                            
#>                  _id trialInformation.globalEndOfTrialPremature
#> 1           12949496                                         NA
#> 2           13281214                                         NA
#> 3           17473621                                         NA
#> 4  2012-003632-23-CZ                                      FALSE
#> 5  2012-003632-23-SE                                      FALSE
#> 6  2014-002606-20-PT                                         NA
#> 7  2014-003556-31-GB                                      FALSE
#> 8  2014-003556-31-SE                                      FALSE
#> 9  2022-501142-30-00                                         NA
#> 10 2023-505613-24-00                                         NA
#> 11 2024-510663-34-00                                         NA
#> 12          20343063                                         NA
#> 13          61070850                                         NA
#> 14          76463425                                         NA
#> 15          80181452                                         NA
#> 16          88261002                                         NA
#> 17       NCT00617929                                         NA
#> 18       NCT01125800                                         NA
#> 19       NCT01483820                                         NA
#> 20       NCT01505608                                         NA
#> 21       NCT01592045                                         NA
#> 22       NCT02620761                                         NA
#> 23       NCT03280147                                         NA
#> 24       NCT03325439                                         NA
#> 25       NCT03431558                                         NA
#> 26       NCT03876704                                         NA
#> 27       NCT04001712                                         NA
#> 28       NCT04041765                                         NA
#> 29       NCT05969327                                         NA
#>    trialInformation.isGlobalEndOfTrialReached p_end_of_trial_status
#> 1                                          NA                  <NA>
#> 2                                          NA                  <NA>
#> 3                                          NA                  <NA>
#> 4                                        TRUE             Completed
#> 5                                        TRUE             Completed
#> 6                                          NA               Ongoing
#> 7                                        TRUE                  <NA>
#> 8                                        TRUE             Completed
#> 9                                          NA                  <NA>
#> 10                                         NA                  <NA>
#> 11                                         NA                  <NA>
#> 12                                         NA                  <NA>
#> 13                                         NA                  <NA>
#> 14                                         NA                  <NA>
#> 15                                         NA                  <NA>
#> 16                                         NA                  <NA>
#> 17                                         NA                  <NA>
#> 18                                         NA                  <NA>
#> 19                                         NA                  <NA>
#> 20                                         NA                  <NA>
#> 21                                         NA                  <NA>
#> 22                                         NA                  <NA>
#> 23                                         NA                  <NA>
#> 24                                         NA                  <NA>
#> 25                                         NA                  <NA>
#> 26                                         NA                  <NA>
#> 27                                         NA                  <NA>
#> 28                                         NA                  <NA>
#> 29                                         NA                  <NA>
#>     last_known_status overall_status protocolSection.statusModule.overallStatus
#> 1                <NA>           <NA>                                       <NA>
#> 2                <NA>           <NA>                                       <NA>
#> 3                <NA>           <NA>                                       <NA>
#> 4                <NA>           <NA>                                       <NA>
#> 5                <NA>           <NA>                                       <NA>
#> 6                <NA>           <NA>                                       <NA>
#> 7                <NA>           <NA>                                       <NA>
#> 8                <NA>           <NA>                                       <NA>
#> 9                <NA>           <NA>                                       <NA>
#> 10               <NA>           <NA>                                       <NA>
#> 11               <NA>           <NA>                                       <NA>
#> 12               <NA>           <NA>                                       <NA>
#> 13               <NA>           <NA>                                       <NA>
#> 14               <NA>           <NA>                                       <NA>
#> 15               <NA>           <NA>                                       <NA>
#> 16               <NA>           <NA>                                       <NA>
#> 17               <NA>           <NA>                                 TERMINATED
#> 18               <NA>           <NA>                                  COMPLETED
#> 19               <NA>           <NA>                                 TERMINATED
#> 20               <NA>           <NA>                                 TERMINATED
#> 21               <NA>           <NA>                                  COMPLETED
#> 22               <NA>     Terminated                                       <NA>
#> 23               <NA>      Completed                                       <NA>
#> 24               <NA>     Terminated                                       <NA>
#> 25         Recruiting Unknown status                                       <NA>
#> 26         Recruiting Unknown status                                       <NA>
#> 27               <NA>      Completed                                       <NA>
#> 28 Not yet recruiting Unknown status                                       <NA>
#> 29               <NA>      Completed                                       <NA>
#>    participants.recruitmentEnd participants.recruitmentStart
#> 1                   2020-01-01                    2015-11-01
#> 2                   2020-01-01                    2016-09-01
#> 3                   2024-02-01                    2023-05-01
#> 4                         <NA>                          <NA>
#> 5                         <NA>                          <NA>
#> 6                         <NA>                          <NA>
#> 7                         <NA>                          <NA>
#> 8                         <NA>                          <NA>
#> 9                         <NA>                          <NA>
#> 10                        <NA>                          <NA>
#> 11                        <NA>                          <NA>
#> 12                  2013-12-31                    2011-01-01
#> 13                  2011-12-31                    2007-01-10
#> 14                  2015-05-31                    2013-06-01
#> 15                  2012-03-31                    2010-04-01
#> 16                  2017-09-28                    2013-09-01
#> 17                        <NA>                          <NA>
#> 18                        <NA>                          <NA>
#> 19                        <NA>                          <NA>
#> 20                        <NA>                          <NA>
#> 21                        <NA>                          <NA>
#> 22                        <NA>                          <NA>
#> 23                        <NA>                          <NA>
#> 24                        <NA>                          <NA>
#> 25                        <NA>                          <NA>
#> 26                        <NA>                          <NA>
#> 27                        <NA>                          <NA>
#> 28                        <NA>                          <NA>
#> 29                        <NA>                          <NA>
#>    participants.recruitmentStatusOverride ctPublicStatusCode   ctStatus
#> 1                                 Stopped                 NA       <NA>
#> 2                                    <NA>                 NA       <NA>
#> 3                                    <NA>                 NA       <NA>
#> 4                                    <NA>                 NA       <NA>
#> 5                                    <NA>                 NA       <NA>
#> 6                                    <NA>                 NA       <NA>
#> 7                                    <NA>                 NA       <NA>
#> 8                                    <NA>                 NA       <NA>
#> 9                                    <NA>                  4 Authorised
#> 10                                   <NA>                  4 Authorised
#> 11                                   <NA>                  4 Authorised
#> 12                                   <NA>                 NA       <NA>
#> 13                                   <NA>                 NA       <NA>
#> 14                                   <NA>                 NA       <NA>
#> 15                                   <NA>                 NA       <NA>
#> 16                                   <NA>                 NA       <NA>
#> 17                                   <NA>                 NA       <NA>
#> 18                                   <NA>                 NA       <NA>
#> 19                                   <NA>                 NA       <NA>
#> 20                                   <NA>                 NA       <NA>
#> 21                                   <NA>                 NA       <NA>
#> 22                                   <NA>                 NA       <NA>
#> 23                                   <NA>                 NA       <NA>
#> 24                                   <NA>                 NA       <NA>
#> 25                                   <NA>                 NA       <NA>
#> 26                                   <NA>                 NA       <NA>
#> 27                                   <NA>                 NA       <NA>
#> 28                                   <NA>                 NA       <NA>
#> 29                                   <NA>                 NA       <NA>
#>    .statusRecruitment
#> 1               other
#> 2           completed
#> 3           completed
#> 4           completed
#> 5           completed
#> 6             ongoing
#> 7           completed
#> 8           completed
#> 9             ongoing
#> 10            ongoing
#> 11            ongoing
#> 12          completed
#> 13          completed
#> 14          completed
#> 15          completed
#> 16          completed
#> 17              other
#> 18          completed
#> 19              other
#> 20              other
#> 21          completed
#> 22              other
#> 23          completed
#> 24              other
#> 25            ongoing
#> 26            ongoing
#> 27          completed
#> 28            ongoing
#> 29          completed

# or use already when creating a trial data frame
# to subset to the unique records of clinical trials
trialsDf <- dbGetFieldsIntoDf(
  calculate = ".isUniqueTrial",
  con = dbc)
#> Querying database (23 fields)...
#> Calculating .isUniqueTrial...                            
#> - Finding duplicates among registers' and sponsor ids...
#> - 2 EUCTR _id were not preferred EU Member State record for 8 trials
#> - Keeping 5 / 3 / 8 / 8 / 3 records from CTGOV2 / EUCTR / CTGOV / ISRCTN / CTIS
trialsDf[trialsDf[[".isUniqueTrial"]], ]
#>                  _id .isUniqueTrial
#> 1           12949496           TRUE
#> 2           13281214           TRUE
#> 3           17473621           TRUE
#> 4  2012-003632-23-CZ           TRUE
#> 6  2014-002606-20-PT           TRUE
#> 8  2014-003556-31-SE           TRUE
#> 9  2022-501142-30-00           TRUE
#> 10 2023-505613-24-00           TRUE
#> 11 2024-510663-34-00           TRUE
#> 12          20343063           TRUE
#> 13          61070850           TRUE
#> 14          76463425           TRUE
#> 15          80181452           TRUE
#> 16          88261002           TRUE
#> 17       NCT00617929           TRUE
#> 18       NCT01125800           TRUE
#> 19       NCT01483820           TRUE
#> 20       NCT01505608           TRUE
#> 21       NCT01592045           TRUE
#> 22       NCT02620761           TRUE
#> 23       NCT03280147           TRUE
#> 24       NCT03325439           TRUE
#> 25       NCT03431558           TRUE
#> 26       NCT03876704           TRUE
#> 27       NCT04001712           TRUE
#> 28       NCT04041765           TRUE
#> 29       NCT05969327           TRUE