Get value for variable of interest

Get information of interest (e.g., endpoint) from long data frame of protocol- or result-related trial information as returned by dfTrials2Long. Parameters `valuename`, `wherename` and `wherevalue` are matched using Perl regular expressions and ignoring case.

Usage

dfName2Value(df, valuename = "", wherename = "", wherevalue = "")

Arguments

df

A data frame (or tibble) with four columns (`_id`, `identifier`, `name`, `value`) as returned by dfTrials2Long

valuename

A character string for the name of the field that holds the value for the variable of interest (e.g., a summary measure such as "endPoints.*tendencyValue.value")

wherename

A character string to identify the variable of interest (e.g., "endPoints.endPoint.title")

wherevalue

A character string with the value of interest for the variable of interest (e.g., "duration of response")

Value

A data frame (or tibble, if dplyr is loaded) with columns `_id`, `identifier`, `name`, `value` that only includes the values of interest, where values are strings unless all value elements are numbers. The `ìdentifier` is generated by function

dfTrials2Long to identify matching elements, e.g. endpoint descriptions and measurements.

Examples

dbc <- nodbi::src_sqlite(
   dbname = system.file("extdata", "demo.sqlite", package = "ctrdata"),
   collection = "my_trials")

df <- dbGetFieldsIntoDf(
fields = c(
  # ctgov - typical results fields
  "clinical_results.baseline.analyzed_list.analyzed.count_list.count",
  "clinical_results.baseline.group_list.group",
  "clinical_results.baseline.analyzed_list.analyzed.units",
  "clinical_results.outcome_list.outcome",
  "study_design_info.allocation",
  # euctr - typical results fields
  "trialInformation.fullTitle",
  "subjectDisposition.recruitmentDetails",
  "baselineCharacteristics.baselineReportingGroups.baselineReportingGroup",
  "endPoints.endPoint",
  "trialChanges.hasGlobalInterruptions",
  "subjectAnalysisSets",
  "adverseEvents.seriousAdverseEvents.seriousAdverseEvent"
  ), con = dbc)
#> clinical_results.baseline.analyzed_list.analyzed.count_list.count... 
#>     
#> . 
#> . 
#> . 
#> . 
#> . 
#> 
#> clinical_results.baseline.group_list.group... 
#>     
#> . 
#> . 
#> . 
#> 
#> clinical_results.baseline.analyzed_list.analyzed.units... 
#>     
#> . 
#> . 
#> . 
#> . 
#> 
#> clinical_results.outcome_list.outcome... 
#>     
#> . 
#> . 
#> 
#> study_design_info.allocation... 
#>     
#> . 
#> 
#> trialInformation.fullTitle... 
#>     
#> . 
#> 
#> subjectDisposition.recruitmentDetails... 
#>     
#> . 
#> 
#> baselineCharacteristics.baselineReportingGroups.baselineReportingGroup... 
#>     
#> . 
#> . 
#> 
#> endPoints.endPoint... 
#>     
#> . 
#> 
#> trialChanges.hasGlobalInterruptions... 
#>     
#> . 
#> 
#> subjectAnalysisSets... 
#>     
#> 
#> adverseEvents.seriousAdverseEvents.seriousAdverseEvent... 
#>     
#> . 
#> . 

# convert to long
reslong <- dfTrials2Long(df = df)
#> clinical_results.baseline.analyzed_list.analyzed.count_list.count                                                                                                                                       
#> clinical_results.baseline.group_list.group                                                                                                                                                              
#> clinical_results.baseline.analyzed_list.analyzed.units                                                                                                                                                  
#> clinical_results.outcome_list.outcome                                                                                                                                                                   
#> study_design_info.allocation                                                                                                                                                                            
#> trialInformation.fullTitle                                                                                                                                                                              
#> subjectDisposition.recruitmentDetails                                                                                                                                                                   
#> baselineCharacteristics.baselineReportingGroups.baselineReportingGroup                                                                                                                                  
#> endPoints.endPoint                                                                                                                                                                                      
#> trialChanges.hasGlobalInterruptions                                                                                                                                                                     
#> subjectAnalysisSets.subjectAnalysisSet                                                                                                                                                                  
#> adverseEvents.seriousAdverseEvents.seriousAdverseEvent                                                                                                                                                  
#>                                                                                                                                                                                                         
#> . 
#> . 
#> . 
#> . 
#> . 
#> . 
#> . 
#> . 
#> . 
#> . 
#> . 
#> . 
#> . 
#> . 
#> . 
#> . 
#> 
#> Total 81865 rows, 131 unique names of variables

# get values for endpoint of interest, e.g. 'response'
dfName2Value(
  df = reslong,
  valuename = paste0(
    "endPoints.endPoint.*armReportingGroup.tendencyValues.tendencyValue.value|",
    "clinical_results.*category.measurement_list.measurement.value|",
    "clinical_results.*outcome.measure.units|endPoints.endPoint.unit"
  ),
  wherename = paste0(
    "clinical_results.*outcome.measure.title|",
    "endPoints.endPoint.title"),
  wherevalue = "response")
#> Returning values for 28 out of 102 trials
#>                   _id identifier
#> 1   2012-003632-23-AT        1.1
#> 2   2012-003632-23-AT        1.2
#> 3   2012-003632-23-AT        1.6
#> 4   2012-003632-23-AT        1.8
#> 5   2012-003632-23-CZ        1.1
#> 6   2012-003632-23-CZ        1.2
#> 7   2012-003632-23-CZ        1.6
#> 8   2012-003632-23-CZ        1.8
#> 9   2012-003632-23-DE        1.1
#> 10  2012-003632-23-DE        1.2
#> 11  2012-003632-23-DE        1.6
#> 12  2012-003632-23-DE        1.8
#> 13  2012-003632-23-ES        1.1
#> 14  2012-003632-23-ES        1.2
#> 15  2012-003632-23-ES        1.6
#> 16  2012-003632-23-ES        1.8
#> 17  2012-003632-23-GB        1.1
#> 18  2012-003632-23-GB        1.2
#> 19  2012-003632-23-GB        1.6
#> 20  2012-003632-23-GB        1.8
#> 21  2012-003632-23-IT        1.1
#> 22  2012-003632-23-IT        1.2
#> 23  2012-003632-23-IT        1.6
#> 24  2012-003632-23-IT        1.8
#> 25  2012-003632-23-SE        1.1
#> 26  2012-003632-23-SE        1.2
#> 27  2012-003632-23-SE        1.6
#> 28  2012-003632-23-SE        1.8
#> 29  2013-003629-27-BE        1.4
#> 30  2013-003629-27-DE        1.4
#> 31  2013-003629-27-DK        1.4
#> 32  2013-003629-27-ES        1.4
#> 33  2013-003629-27-FR        1.4
#> 34  2013-003629-27-GB        1.4
#> 35  2013-003629-27-IE        1.4
#> 36  2013-003629-27-IT        1.4
#> 37  2013-003629-27-NL        1.4
#> 38  2013-003629-27-PL        1.4
#> 39  2013-003629-27-SE        1.4
#> 40  2013-005512-10-DE        1.4
#> 41  2013-005512-10-DE      1.4.1
#> 42  2013-005512-10-DE      1.4.2
#> 43  2013-005512-10-DE      1.4.3
#> 44  2013-005512-10-DE      1.4.4
#> 45  2013-005512-10-DE      1.4.5
#> 46  2013-005512-10-DE      1.4.6
#> 47  2013-005512-10-DE      1.4.7
#> 48  2013-005512-10-DE      1.4.8
#> 49  2013-005512-10-DE      1.4.9
#> 50  2013-005512-10-DE     1.4.10
#> 51  2013-005512-10-DE     1.4.11
#> 52  2013-005512-10-DE     1.4.12
#> 53  2013-005512-10-DE       1.10
#> 54  2013-005512-10-DK        1.4
#> 55  2013-005512-10-DK      1.4.1
#> 56  2013-005512-10-DK      1.4.2
#> 57  2013-005512-10-DK      1.4.3
#> 58  2013-005512-10-DK      1.4.4
#> 59  2013-005512-10-DK      1.4.5
#> 60  2013-005512-10-DK      1.4.6
#> 61  2013-005512-10-DK      1.4.7
#> 62  2013-005512-10-DK      1.4.8
#> 63  2013-005512-10-DK      1.4.9
#> 64  2013-005512-10-DK     1.4.10
#> 65  2013-005512-10-DK     1.4.11
#> 66  2013-005512-10-DK     1.4.12
#> 67  2013-005512-10-DK       1.10
#> 68  2013-005512-10-ES        1.4
#> 69  2013-005512-10-ES      1.4.1
#> 70  2013-005512-10-ES      1.4.2
#> 71  2013-005512-10-ES      1.4.3
#> 72  2013-005512-10-ES      1.4.4
#> 73  2013-005512-10-ES      1.4.5
#> 74  2013-005512-10-ES      1.4.6
#> 75  2013-005512-10-ES      1.4.7
#> 76  2013-005512-10-ES      1.4.8
#> 77  2013-005512-10-ES      1.4.9
#> 78  2013-005512-10-ES     1.4.10
#> 79  2013-005512-10-ES     1.4.11
#> 80  2013-005512-10-ES     1.4.12
#> 81  2013-005512-10-ES       1.10
#> 82  2013-005512-10-FR        1.4
#> 83  2013-005512-10-FR      1.4.1
#> 84  2013-005512-10-FR      1.4.2
#> 85  2013-005512-10-FR      1.4.3
#> 86  2013-005512-10-FR      1.4.4
#> 87  2013-005512-10-FR      1.4.5
#> 88  2013-005512-10-FR      1.4.6
#> 89  2013-005512-10-FR      1.4.7
#> 90  2013-005512-10-FR      1.4.8
#> 91  2013-005512-10-FR      1.4.9
#> 92  2013-005512-10-FR     1.4.10
#> 93  2013-005512-10-FR     1.4.11
#> 94  2013-005512-10-FR     1.4.12
#> 95  2013-005512-10-FR       1.10
#> 96  2013-005512-10-GB        1.4
#> 97  2013-005512-10-GB      1.4.1
#> 98  2013-005512-10-GB      1.4.2
#> 99  2013-005512-10-GB      1.4.3
#> 100 2013-005512-10-GB      1.4.4
#> 101 2013-005512-10-GB      1.4.5
#> 102 2013-005512-10-GB      1.4.6
#> 103 2013-005512-10-GB      1.4.7
#> 104 2013-005512-10-GB      1.4.8
#> 105 2013-005512-10-GB      1.4.9
#> 106 2013-005512-10-GB     1.4.10
#> 107 2013-005512-10-GB     1.4.11
#> 108 2013-005512-10-GB     1.4.12
#> 109 2013-005512-10-GB       1.10
#> 110 2013-005512-10-IE        1.4
#> 111 2013-005512-10-IE      1.4.1
#> 112 2013-005512-10-IE      1.4.2
#> 113 2013-005512-10-IE      1.4.3
#> 114 2013-005512-10-IE      1.4.4
#> 115 2013-005512-10-IE      1.4.5
#> 116 2013-005512-10-IE      1.4.6
#> 117 2013-005512-10-IE      1.4.7
#> 118 2013-005512-10-IE      1.4.8
#> 119 2013-005512-10-IE      1.4.9
#> 120 2013-005512-10-IE     1.4.10
#> 121 2013-005512-10-IE     1.4.11
#> 122 2013-005512-10-IE     1.4.12
#> 123 2013-005512-10-IE       1.10
#> 124 2013-005512-10-IT        1.4
#> 125 2013-005512-10-IT      1.4.1
#> 126 2013-005512-10-IT      1.4.2
#> 127 2013-005512-10-IT      1.4.3
#> 128 2013-005512-10-IT      1.4.4
#> 129 2013-005512-10-IT      1.4.5
#> 130 2013-005512-10-IT      1.4.6
#> 131 2013-005512-10-IT      1.4.7
#> 132 2013-005512-10-IT      1.4.8
#> 133 2013-005512-10-IT      1.4.9
#> 134 2013-005512-10-IT     1.4.10
#> 135 2013-005512-10-IT     1.4.11
#> 136 2013-005512-10-IT     1.4.12
#> 137 2013-005512-10-IT       1.10
#> 138 2013-005512-10-NL        1.4
#> 139 2013-005512-10-NL      1.4.1
#> 140 2013-005512-10-NL      1.4.2
#> 141 2013-005512-10-NL      1.4.3
#> 142 2013-005512-10-NL      1.4.4
#> 143 2013-005512-10-NL      1.4.5
#> 144 2013-005512-10-NL      1.4.6
#> 145 2013-005512-10-NL      1.4.7
#> 146 2013-005512-10-NL      1.4.8
#> 147 2013-005512-10-NL      1.4.9
#> 148 2013-005512-10-NL     1.4.10
#> 149 2013-005512-10-NL     1.4.11
#> 150 2013-005512-10-NL     1.4.12
#> 151 2013-005512-10-NL       1.10
#> 152 2013-005512-10-PL        1.4
#> 153 2013-005512-10-PL      1.4.1
#> 154 2013-005512-10-PL      1.4.2
#> 155 2013-005512-10-PL      1.4.3
#> 156 2013-005512-10-PL      1.4.4
#> 157 2013-005512-10-PL      1.4.5
#> 158 2013-005512-10-PL      1.4.6
#> 159 2013-005512-10-PL      1.4.7
#> 160 2013-005512-10-PL      1.4.8
#> 161 2013-005512-10-PL      1.4.9
#> 162 2013-005512-10-PL     1.4.10
#> 163 2013-005512-10-PL     1.4.11
#> 164 2013-005512-10-PL     1.4.12
#> 165 2013-005512-10-PL       1.10
#> 166 2013-005512-10-SE        1.4
#> 167 2013-005512-10-SE      1.4.1
#> 168 2013-005512-10-SE      1.4.2
#> 169 2013-005512-10-SE      1.4.3
#> 170 2013-005512-10-SE      1.4.4
#> 171 2013-005512-10-SE      1.4.5
#> 172 2013-005512-10-SE      1.4.6
#> 173 2013-005512-10-SE      1.4.7
#> 174 2013-005512-10-SE      1.4.8
#> 175 2013-005512-10-SE      1.4.9
#> 176 2013-005512-10-SE     1.4.10
#> 177 2013-005512-10-SE     1.4.11
#> 178 2013-005512-10-SE     1.4.12
#> 179 2013-005512-10-SE       1.10
#>                                                                                           name
#> 1                                                                      endPoints.endPoint.unit
#> 2                                                                      endPoints.endPoint.unit
#> 3                                                                      endPoints.endPoint.unit
#> 4                                                                      endPoints.endPoint.unit
#> 5                                                                      endPoints.endPoint.unit
#> 6                                                                      endPoints.endPoint.unit
#> 7                                                                      endPoints.endPoint.unit
#> 8                                                                      endPoints.endPoint.unit
#> 9                                                                      endPoints.endPoint.unit
#> 10                                                                     endPoints.endPoint.unit
#> 11                                                                     endPoints.endPoint.unit
#> 12                                                                     endPoints.endPoint.unit
#> 13                                                                     endPoints.endPoint.unit
#> 14                                                                     endPoints.endPoint.unit
#> 15                                                                     endPoints.endPoint.unit
#> 16                                                                     endPoints.endPoint.unit
#> 17                                                                     endPoints.endPoint.unit
#> 18                                                                     endPoints.endPoint.unit
#> 19                                                                     endPoints.endPoint.unit
#> 20                                                                     endPoints.endPoint.unit
#> 21                                                                     endPoints.endPoint.unit
#> 22                                                                     endPoints.endPoint.unit
#> 23                                                                     endPoints.endPoint.unit
#> 24                                                                     endPoints.endPoint.unit
#> 25                                                                     endPoints.endPoint.unit
#> 26                                                                     endPoints.endPoint.unit
#> 27                                                                     endPoints.endPoint.unit
#> 28                                                                     endPoints.endPoint.unit
#> 29                                                                     endPoints.endPoint.unit
#> 30                                                                     endPoints.endPoint.unit
#> 31                                                                     endPoints.endPoint.unit
#> 32                                                                     endPoints.endPoint.unit
#> 33                                                                     endPoints.endPoint.unit
#> 34                                                                     endPoints.endPoint.unit
#> 35                                                                     endPoints.endPoint.unit
#> 36                                                                     endPoints.endPoint.unit
#> 37                                                                     endPoints.endPoint.unit
#> 38                                                                     endPoints.endPoint.unit
#> 39                                                                     endPoints.endPoint.unit
#> 40                                                                     endPoints.endPoint.unit
#> 41  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 42  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 43  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 44  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 45  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 46  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 47  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 48  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 49  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 50  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 51  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 52  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 53                                                                     endPoints.endPoint.unit
#> 54                                                                     endPoints.endPoint.unit
#> 55  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 56  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 57  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 58  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 59  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 60  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 61  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 62  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 63  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 64  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 65  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 66  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 67                                                                     endPoints.endPoint.unit
#> 68                                                                     endPoints.endPoint.unit
#> 69  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 70  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 71  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 72  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 73  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 74  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 75  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 76  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 77  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 78  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 79  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 80  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 81                                                                     endPoints.endPoint.unit
#> 82                                                                     endPoints.endPoint.unit
#> 83  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 84  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 85  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 86  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 87  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 88  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 89  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 90  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 91  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 92  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 93  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 94  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 95                                                                     endPoints.endPoint.unit
#> 96                                                                     endPoints.endPoint.unit
#> 97  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 98  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 99  endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 100 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 101 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 102 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 103 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 104 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 105 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 106 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 107 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 108 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 109                                                                    endPoints.endPoint.unit
#> 110                                                                    endPoints.endPoint.unit
#> 111 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 112 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 113 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 114 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 115 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 116 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 117 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 118 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 119 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 120 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 121 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 122 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 123                                                                    endPoints.endPoint.unit
#> 124                                                                    endPoints.endPoint.unit
#> 125 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 126 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 127 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 128 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 129 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 130 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 131 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 132 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 133 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 134 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 135 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 136 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 137                                                                    endPoints.endPoint.unit
#> 138                                                                    endPoints.endPoint.unit
#> 139 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 140 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 141 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 142 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 143 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 144 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 145 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 146 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 147 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 148 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 149 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 150 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 151                                                                    endPoints.endPoint.unit
#> 152                                                                    endPoints.endPoint.unit
#> 153 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 154 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 155 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 156 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 157 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 158 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 159 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 160 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 161 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 162 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 163 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 164 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 165                                                                    endPoints.endPoint.unit
#> 166                                                                    endPoints.endPoint.unit
#> 167 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 168 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 169 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 170 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 171 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 172 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 173 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 174 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 175 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 176 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 177 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 178 endPoints.endPoint.armReportingGroups.armReportingGroup.tendencyValues.tendencyValue.value
#> 179                                                                    endPoints.endPoint.unit
#>                            value
#> 1                           Days
#> 2   At least 1 response (number)
#> 3          Overall Response Rate
#> 4   Percentage of treatment time
#> 5                           Days
#> 6   At least 1 response (number)
#> 7          Overall Response Rate
#> 8   Percentage of treatment time
#> 9                           Days
#> 10  At least 1 response (number)
#> 11         Overall Response Rate
#> 12  Percentage of treatment time
#> 13                          Days
#> 14  At least 1 response (number)
#> 15         Overall Response Rate
#> 16  Percentage of treatment time
#> 17                          Days
#> 18  At least 1 response (number)
#> 19         Overall Response Rate
#> 20  Percentage of treatment time
#> 21                          Days
#> 22  At least 1 response (number)
#> 23         Overall Response Rate
#> 24  Percentage of treatment time
#> 25                          Days
#> 26  At least 1 response (number)
#> 27         Overall Response Rate
#> 28  Percentage of treatment time
#> 29       responses to injections
#> 30       responses to injections
#> 31       responses to injections
#> 32       responses to injections
#> 33       responses to injections
#> 34       responses to injections
#> 35       responses to injections
#> 36       responses to injections
#> 37       responses to injections
#> 38       responses to injections
#> 39       responses to injections
#> 40       responses to injections
#> 41                           102
#> 42                            16
#> 43                             2
#> 44                           118
#> 45                           163
#> 46                            27
#> 47                            14
#> 48                            89
#> 49                            20
#> 50                            14
#> 51                             3
#> 52                            11
#> 53                      subjects
#> 54       responses to injections
#> 55                           102
#> 56                            16
#> 57                             2
#> 58                           118
#> 59                           163
#> 60                            27
#> 61                            14
#> 62                            89
#> 63                            20
#> 64                            14
#> 65                             3
#> 66                            11
#> 67                      subjects
#> 68       responses to injections
#> 69                           102
#> 70                            16
#> 71                             2
#> 72                           118
#> 73                           163
#> 74                            27
#> 75                            14
#> 76                            89
#> 77                            20
#> 78                            14
#> 79                             3
#> 80                            11
#> 81                      subjects
#> 82       responses to injections
#> 83                           102
#> 84                            16
#> 85                             2
#> 86                           118
#> 87                           163
#> 88                            27
#> 89                            14
#> 90                            89
#> 91                            20
#> 92                            14
#> 93                             3
#> 94                            11
#> 95                      subjects
#> 96       responses to injections
#> 97                           102
#> 98                            16
#> 99                             2
#> 100                          118
#> 101                          163
#> 102                           27
#> 103                           14
#> 104                           89
#> 105                           20
#> 106                           14
#> 107                            3
#> 108                           11
#> 109                     subjects
#> 110      responses to injections
#> 111                          102
#> 112                           16
#> 113                            2
#> 114                          118
#> 115                          163
#> 116                           27
#> 117                           14
#> 118                           89
#> 119                           20
#> 120                           14
#> 121                            3
#> 122                           11
#> 123                     subjects
#> 124      responses to injections
#> 125                          102
#> 126                           16
#> 127                            2
#> 128                          118
#> 129                          163
#> 130                           27
#> 131                           14
#> 132                           89
#> 133                           20
#> 134                           14
#> 135                            3
#> 136                           11
#> 137                     subjects
#> 138      responses to injections
#> 139                          102
#> 140                           16
#> 141                            2
#> 142                          118
#> 143                          163
#> 144                           27
#> 145                           14
#> 146                           89
#> 147                           20
#> 148                           14
#> 149                            3
#> 150                           11
#> 151                     subjects
#> 152      responses to injections
#> 153                          102
#> 154                           16
#> 155                            2
#> 156                          118
#> 157                          163
#> 158                           27
#> 159                           14
#> 160                           89
#> 161                           20
#> 162                           14
#> 163                            3
#> 164                           11
#> 165                     subjects
#> 166      responses to injections
#> 167                          102
#> 168                           16
#> 169                            2
#> 170                          118
#> 171                          163
#> 172                           27
#> 173                           14
#> 174                           89
#> 175                           20
#> 176                           14
#> 177                            3
#> 178                           11
#> 179                     subjects