Calculate the external references from a study's register record

Trial concept calculated: Calculates the links e.g. to publications or other external files referenced from a study record. Requires loading results-related information for EUCTR. Note that documents stored in registers can be downloaded directly, see ctrLoadQueryIntoDb.

Usage

f.externalLinks(df = NULL)

Arguments

df

data frame such as from dbGetFieldsIntoDf. If `NULL`, prints fields needed in `df` for calculating this trial concept, which can be used with dbGetFieldsIntoDf.

Value

data frame with columns `_id` and new column `.externalLinks` (character).

Examples

# fields needed
f.hasResults()
#> $euctr
#> [1] "endPoints.endPoint.readyForValues"
#> 
#> $ctgov
#> [1] "results_reference.citation"                
#> [2] "clinical_results.outcome_list.outcome.type"
#> 
#> $ctgov2
#> [1] "hasResults"                                               
#> [2] "protocolSection.referencesModule.references.type"         
#> [3] "protocolSection.statusModule.resultsFirstSubmitDate"      
#> [4] "resultsSection.outcomeMeasuresModule.outcomeMeasures.type"
#> 
#> $isrctn
#> [1] "results.publicationStage"
#> 
#> $ctis
#> [1] "results.clinicalStudyReports.id" "results.laypersonResults.id"    
#> [3] "results.summaryResults.id"       "resultsFirstReceived"           
#> 

# apply trial concept when creating data frame
dbc <- nodbi::src_sqlite(
  dbname = system.file("extdata", "demo.sqlite", package = "ctrdata"),
  collection = "my_trials", flags = RSQLite::SQLITE_RO)
trialsDf <- dbGetFieldsIntoDf(
  calculate = "f.externalLinks",
  con = dbc)
#> To review trial concepts details, call 'help("ctrdata-trial-concepts")'
#> Querying database (9 fields)...
#> Calculating f.externalLinks...                            
trialsDf
#>            _id
#> 1     12949496
#> 2     13281214
#> 3     76463425
#> 4     80181452
#> 5     88261002
#> 6  NCT01483820
#> 7  NCT01505608
#> 8  NCT03325439
#> 9  NCT03431558
#> 10 NCT03876704
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         .externalLinks
#> 1                                                                                                                                                                                                                                                                                                                                                                   2018 Protocol article in https://www.ncbi.nlm.nih.gov/pubmed/30524749 protocol2022 Results article in https://pubmed.ncbi.nlm.nih.gov/36385642/ (added 18/11/2022)
#> 2                                                                                                                                            2021 Results article in https://pubmed.ncbi.nlm.nih.gov/33069668/ results (added 20/10/2020)2021 Other publications in https://pubmed.ncbi.nlm.nih.gov/33310351/ cohort study (added 16/12/2020)2020 Protocol article in https://pubmed.ncbi.nlm.nih.gov/34522836/ RCT protocol (added 21/09/2021)2022 Other publications in https://pubmed.ncbi.nlm.nih.gov/35585179/ (added 18/08/2023)
#> 3                                                                                                                                             2017 Protocol article in https://www.ncbi.nlm.nih.gov/pubmed/28129748 protocol2019 Results article in https://www.ncbi.nlm.nih.gov/pubmed/31597020 results (added 10/10/2019)2020 Results article in https://pubmed.ncbi.nlm.nih.gov/32342857 results (added 29/04/2020)2021 Results article in https://pubmed.ncbi.nlm.nih.gov/34419121/ Study Within A Trial (SWAT) (added 23/08/2021)
#> 4                                                                                                                                                                                                                                                                                  2010 Protocol article in http://www.ncbi.nlm.nih.gov/pubmed/19590247 protocol2015 Results article in https://www.nejm.org/doi/full/10.1056/NEJMoa1501917 results2018 Results article in https://www.nejm.org/doi/full/10.1056/NEJMoa1708831 results
#> 5                                                                                                                                                                                                                                                                                                                                                                             2018 Results article in https://www.ncbi.nlm.nih.gov/pubmed/30574860 results2019 Results article in https://www.ncbi.nlm.nih.gov/pubmed/30635141 results
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                               Beat Childhood Cancer Consortium http://www.beatcc.org
#> 7                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            Beat Childhood Cancer https://beatcc.org/
#> 8  Pressler R, Boylan G, Dempsey E, Klotz KA, Krauwinkel W, Will E, Morita D, Floricel F, Elshoff JP, van den Anker J. Pharmacokinetics and safety of brivaracetam in neonates with repeated electroencephalographic seizures: A multicenter, open-label, single-arm study. Epilepsia Open. 2024 Apr;9(2):522-533. doi: 10.1002/epi4.12875. Epub 2024 Jan 11. https://www.briviact.com/briviact-PI.pdf?v=1479491757 http://www.fda.gov/Safety/MedWatch/SafetyInformation/default.htm Product Information FDA Safety Alerts and Recalls
#> 9                                                                                                                                                                                                                                                                                                               http://data.worldbank.org/indicator/SH.DYN.NMRT http://www.un.org/sustainabledevelopment/health The World Bank Mortality rate ,Neonatal (per,1000 live births) 29/02/2016 United.Nations Sustainable development goals
#> 10                                                                                                                                                                                                                                                       Fares S, Sethom MM, Khouaja-Mokrani C, Jabnoun S, Feki M, Kaabachi N. VitaminA, E, and D deficiencies in tunisian very low birth weight neonates: prevalence and risk factors. Pediatr Neonatol. 2014 Jun;55(3):196-201. doi: 10.1016/j.pedneo.2013.09.006. Epub 2013 Nov 26.