Calculate the title of a study

Trial concept calculated: scientific or full title of the study.

Usage

f.trialTitle(df = NULL)

Arguments

df

data frame such as from dbGetFieldsIntoDf. If `NULL`, prints fields needed in `df` for calculating this trial concept, which can be used with dbGetFieldsIntoDf.

Value

data frame with columns `_id` and `.trialTitle`, a string.

Examples

# fields needed
f.resultsDate()
#> $euctr
#> [1] "firstreceived_results_date"         "trialInformation.analysisStageDate"
#> 
#> $ctgov
#> [1] "results_first_posted"
#> 
#> $ctgov2
#> [1] "protocolSection.statusModule.resultsFirstPostDateStruct.date"
#> 
#> $isrctn
#> [1] "results.intentToPublish"
#> 
#> $ctis
#> [1] "results.summaryResults.submissionDate"  
#> [2] "results.clinicalStudyReports.submitDate"
#> 

# apply trial concept when creating data frame
dbc <- nodbi::src_sqlite(
  dbname = system.file("extdata", "demo.sqlite", package = "ctrdata"),
  collection = "my_trials", flags = RSQLite::SQLITE_RO)
trialsDf <- dbGetFieldsIntoDf(
  calculate = "f.trialTitle",
  con = dbc)
#> Querying database (6 fields)...
trialsDf
#>                  _id
#> 1           12949496
#> 2           13281214
#> 3           17473621
#> 4  2012-003632-23-CZ
#> 5  2012-003632-23-SE
#> 6  2014-002606-20-PT
#> 7  2014-003556-31-GB
#> 8  2014-003556-31-SE
#> 9  2022-500244-37-00
#> 10 2022-501142-30-00
#> 11 2023-505613-24-00
#> 12 2023-508143-51-01
#> 13 2024-510663-34-00
#> 14          20343063
#> 15          61070850
#> 16          76463425
#> 17          80181452
#> 18          88261002
#> 19       NCT00617929
#> 20       NCT01125800
#> 21       NCT01483820
#> 22       NCT01505608
#> 23       NCT01592045
#> 24       NCT02620761
#> 25       NCT03280147
#> 26       NCT03325439
#> 27       NCT03431558
#> 28       NCT03876704
#> 29       NCT04001712
#> 30       NCT04041765
#> 31       NCT05969327
#>                                                                                                                                                                                                                                                                                                                                                                                 .trialTitle
#> 1                                                                                                                                                                                                                                                         The use of milrinone in neonates with persistent pulmonary hypertension of the newborn: a randomised controlled trial pilot study
#> 2                                                                                                                                                                                                                                                                      A randomised controlled trial of early targeted patent ductus arteriosus treatment using a risk-based severity score
#> 3                                                                                                                                                                                     The effectiveness of the “BLUI Blanket” light emitting diode phototherapy blanket model against decreased bilirubin in physiological jaundice gestational age ≥ 35 Weeks - a randomised control trial
#> 4  A Pilot, Open-label, Single Arm, Multicentre Study to Explore Safety, Tolerability, Pharmacokinetics and Efficacy of Intravenous Multiple Administrations of NI-0501, an Anti-interferon Gamma Anti-IFN Monoclonal Antibody, in Paediatric Patients with Primary Haemophagocytic Lymphohistiocytosis in whom the disease has reactivated or an unsatisfactory response has been achieved
#> 5                                                                               A Phase 2/3, Open-label, Single Arm, Multicentre Study to Assess Safety, Tolerability, Pharmacokinetics and Efficacy of Intravenous Multiple Administrations of NI-0501, an Anti-interferon Gamma Anti-IFN Monoclonal Antibody, in Paediatric Patients with Primary Haemophagocytic Lymphohistiocytosis HLH
#> 6                                                                                                                                                                                                     A RANDOMIZED, OPEN-LABEL, ACTIVE CONTROLLED, SAFETY AND DESCRIPTIVE EFFICACY STUDY IN PEDIATRIC SUBJECTS REQUIRING ANTICOAGULATION FOR THE TREATMENT OF A VENOUS THROMBOEMBOLIC EVENT
#> 7                                                                                                                                                                                        Long-term Outcome of Children Enrolled in Study ROPP-2008-01 Previously Treated with rhIGF-1/rhIGFBP-3 for the Prevention of Retinopathy of Prematurity ROP or Who Received Standard Neonatal Care
#> 8                                                                                                                                                                                        Long-term Outcome of Children Enrolled in Study ROPP-2008-01 Previously Treated with rhIGF-1/rhIGFBP-3 for the Prevention of Retinopathy of Prematurity ROP or Who Received Standard Neonatal Care
#> 9                                                                                           A 4-week double-blind, randomized, placebo-controlled, phase II study evaluating the effects of oral pamapimod 150 mg with pioglitazone 10 mg daily on  COVID-19 evolution and recovery in non-hospitalized patients infected with SARS-CoV-2 (severe acute respiratory syndrome-coronavirus-2)
#> 10                                                                                                                                                                                                                                                       Balloon + Oxytocin versus Oral Misoprostol to Induce labor in case of premature rupture of membranes (PROM) at term in nulliparous
#> 11                                                A Phase 2, Open-Label, Single-Arm, Sequential-Panel Study to Evaluate the Pharmacokinetics, Safety, and Tolerability of Posaconazole (POS, MK-5592) Intravenous and Powder for Oral Suspension Formulations in Pediatric Participants From Birth to Less Than 2 Years of Age With Possible, Probable, or Proven Invasive Fungal Infection
#> 12                                                                                                                                                                                                                                                  Pulmonary function and sleep related disorders during cervical admission of intrathecal baclofen in spinal cord injury;\na safety study
#> 13                                                                                              An open label, single arm study to evaluate single and multiple dose pharmacokinetics, safety and tolerability, and to explore clinical outcomes of treatment with intravenous (IV) zanamivir in neonates and infants under 6 months of age with confirmed complicated influenza infection.
#> 14                                                                                                                                                                                                                     Randomised analysis, double blind and controlled with placebo of the utility of the extract of cranberry in the prevention of the enterocolitis of premature infants
#> 15                                                                                                                                            Efficacy and safety of preterm formula addition with medium chain triglycerides or maltodextrins compared with milk formula without addition of energy substrates for nutrition of newborn very low birth weight: a randomised clinical trial
#> 16                                                                                                                                                                                                                                             A multi-centre randomised controlled trial of two speeds of daily increment of milk feeding in very preterm or very low birth weight infants
#> 17                                                                                                                                                                                                                                                              Efficacy and safety of inhaled budesonide in very preterm infants at risk for bronchopulmonary dysplasia: a phase III trial
#> 18                                                                                                                                                                                                                  A multi-centre randomised placebo-controlled trial of prophylactic enteral lactoferrin supplementation to prevent late-onset invasive infection in very preterm infants
#> 19                                                                                                                                                                                                                                                                                                             Conditioning for Graft Failure After Hematopoietic Stem Cell Transplantation
#> 20                                                                                                                                                  A Phase I/II Study of LDE225 in Pediatric Patients With Recurrent or Refractory Medulloblastoma or Other Tumors Potentially Dependent on the Hedgehog-signaling Pathway and Adult Patients With Recurrent or Refractory Medulloblastoma
#> 21                                                                                                                                                                                                                                                                                 A Phase I/II Trial of TPI-287 in Patients With Refractory or Recurrent Neuroblastoma and Medulloblastoma
#> 22                                                                                                                                                                                    A Randomized, Phase I/II Trial of Irinotecan and Temozolomide Compared to Irinotecan and Temozolomide in Combination With TPI 287 in Patients With Primary Refractory or Early Relapsed Neuroblastoma
#> 23                                                                                                                         A Comparative Pharmacokinetic and Safety Study of Chimeric Monoclonal Antibody ch14.18 With Granulocyte-Macrophage Colony-Stimulating Factor (GM-CSF), Interleukin-2 (IL-2) and Isotretinoin in High Risk Neuroblastoma Patients Following Myeloablative Therapy
#> 24                                                                                                                                                                                                                                                                                                          Fenoldopam to Prevent Renal Dysfunction in Indomethacin Treated Preterm Infants
#> 25                                                                                                                                                                                         Comparison of the Efficacy of a 7-day Versus 14-day Course of Intravenous Antibiotics in the Treatment of Uncomplicated Neonatal Bacterial Sepsis: a Randomized Controlled Non-inferiority Trial
#> 26                                                                                                                                                                                                            A Multicenter, Open-Label, Single-Arm Study to Evaluate the Pharmacokinetics, Efficacy, and Safety of Brivaracetam in Neonates With Repeated Electroencephalographic Seizures
#> 27                                                                                                                                                                                                                                                                               Can Bovine Lactoferrin Prevent Neonatal Infections in Low Birth Weight Babies in Karachi, Sindh, Pakistan.
#> 28                                                                                                                                                                                                                                            Effects of Fat-soluble Vitamins Supplementation in Early Life on Common Complications and Neural Development in Very Low Birth Weight Infants
#> 29                                                                                                                                                                                                                                                                                                 Effect of Early Use of Caffeine Citrate in Preterm Neonates Needing Respiratory Support.
#> 30                                                                                                                                                                                                          Efficacy of Prophylactic IgM-Enriched Immunoglobulin for the Management of Early-Onset Neonatal Sepsis in Very Low Birth Weight Preterm Neonates; A Randomized Controlled Trial
#> 31                                                                                                                                                                                                                                                                                           Role of Vitamin D Therapy in Recovery From Early Neonatal Sepsis (Randomized Controlled Trial)