Open stored query or register search page

Open advanced search pages of register(s), or execute search in browser

Usage

ctrOpenSearchPagesInBrowser(url = "", register = "", copyright = FALSE, ...)

Arguments

url

of search results page to show in the browser. Can be the return value of ctrGetQueryUrl or dbQueryHistory.

register

Register(s) to open. Either "EUCTR" or "CTGOV" or a vector of both. Default is to open both registers' advanced search pages. To open the browser with a previous search, the output of ctrGetQueryUrl or one row from dbQueryHistory can be used.

copyright

(Optional) If set to TRUE, opens copyright pages of register(s).

...

May include the deprecated input parameter.

Value

Always TRUE, invisibly.

Examples

# Check copyrights before using registers
ctrOpenSearchPagesInBrowser(copyright = TRUE)

# open last query loaded into the collection
dbc <- nodbi::src_sqlite(
   dbname = system.file("extdata", "demo.sqlite", package = "ctrdata"),
   collection = "my_trials")

ctrOpenSearchPagesInBrowser(
  dbQueryHistory(con = dbc))
#> Warning: Using last query
#> [1] "https://www.isrctn.com/search?q=&filters=ageRange:Neonate,phase:Phase+III"