ctrdata: information on clinical trial registers

Registers of clinical trials that can be accessed with package ctrdata as of September 2023.

Details

EUCTR: The European Union Clinical Trials Register contains more than 43,500 clinical trials (using one or more medicines as investigational medicinal product, IMP; in the European Union and beyond)

CTGOV and CTGOV2: ClinicalTrials.gov includes more than 465,000 interventional and observational studies (both the current and the classic website are supported)

ISRCTN: The ISRCTN Registry includes almost 24,000 interventional or observational health studies

CTIS: The EU Clinical Trials Information System was started in January 2023, including more than 730 publicly accessible trials on 2024-04-27. How to automatically get a query URL: here

Material	EUCTR	CTGOV	CTGOV2	ISRCTN	CTIS

Home page	link	link	link	link

About	link	link	link	link	link

Terms and conditions, disclaimer	link	link	link	link	link

How to search	link	link	link	link	link

Search interface	link	link	link	link	link

Glossary	link	link	link	link

FAQ	link	link	link	link

Expert / advanced search	link	link		link	link

Example*	link	link	link	link	link

Definitions	link	link	link	link	link

*The example is an expert search that retrieves interventional trials with neonates, investigating infectious conditions: The CTGOV expert search retrieves trials conducted exclusively with neonates. It is not yet clear why different sets of trials are shown in CTGOV (65) and CTGOV2 (14) for the same query parameters. EUCTR retrieves trials with neonates, but not only those exclusively in neonates. ISRCTN retrieves studies with interventions other than medicines. Thus, after loading trials with ctrLoadQueryIntoDb into a database collection, trials of interest have to be selected in a second step, based on values of fields of interest (e.g., using f115_children_211years from EUCTR for age criteria, interventions.intervention.interventionType from ISRCTN for type of study), which can be obtained from the collection using dbGetFieldsIntoDf.

Author

Ralf Herold ralf.herold@mailbox.org