Information on clinical trial registers

Registers of the four clinical trial registers from which package ctrdata can retrieve, aggregate and analyse protocol- and result-related information as well as documents, last updated 2025-05-14.

1 - Overview

EUCTR: The EU Clinical Trials Register holds more than 44,300 clinical trials (at least one investigational medicinal product, IMP; in the European Union and beyond), including more than 25,000 trials with results, which continue to be added (and can be loaded by ctrdata).

CTIS: The EU Clinical Trials Information System, launched in 2023, holds more than 9,100 publicly accessible clinical trials, including around 100 with results or a report (only as PDF files). No results in a structured electronic format are foreseeably available, thus ctrdata cannot load any CTIS results. (To automatically get CTIS search query URLs, see here)

CTGOV2: ClinicalTrials.gov holds more than 537,000 interventional and observational studies, including almost 67,000 interventional studies with results (can be loaded by ctrdata).

ISRCTN: The ISRCTN Registry holds more than 26,300 interventional and observational health studies, including almost 14,300 studies with results (only as references). No results in a structured electronic format are foreseeably available, thus ctrdata cannot load any ISRCTN results.

2 - Notable changes

CTGOV "classic" was retired on 2024-06-25; ctrdata subsequently translates CTGOV queries to CTGOV2 queries. The new website ("CTGOV2") can be used with ctrdata since 2023-08-27. Database collections created with CTGOV queries can still be used since functions in ctrdata continue to support them. CTIS was relaunched on 2024-06-17, changing the data structure and search syntax, to which ctrdata was updated. CTIS can be used with ctrdata since 2023-03-25. EUCTR removed search parameter status= as of February 2025. More information on changes: here.

3 - References

Material	EUCTR	CTGOV2	ISRCTN	CTIS

About	link	link	link	link

Terms & conditions, disclaimer	link	link	link	link

How to search	link	link	link	link

Search interface	link	link	link	link

Expert / advanced search	link	link	link	link

Glossary / related information	link	link	link	link

FAQ, caveats, issues	link	link, link	link	link

Data dictionaries / definitions / structure reference	link, link, link	link, link, link	link	link (XLSX files)

Example*	link	link	link	link

Some registers are expanding entered search terms using dictionaries (example).

4 - Example and ctrdata motivation

See vignette("ctrdata_summarise") for several other examples.

*This example is an expert search for interventional trials primarily with neonates, investigating treatments for infectious conditions. It shows that searches in the web interface of most registers are not sufficient to identify the trials of interest:

EUCTR retrieves trials with neonates, but not only those exclusively in neonates.

ISRCTN retrieves studies with interventions other than medicines.

CTIS retrieves trials that mention the words neonates and infection. (To show CTIS search results, see here)

To address this issue, trials can be retrieved with ctrLoadQueryIntoDb into a database collection and in a second step trials of interest can be selected based on values of relevant fields, for example:

EUCTR field f115_children_211years and other age group criteria

ISRCTN field interventions.intervention.interventionType for type of study

CTIS fields ageGroup and authorizedApplication.authorizedPartI.medicalConditions.medicalCondition

ctrdata supports users with pre-defined ctrdata-trial-concepts and these cover the example above, and with functions dbFindFields and ctrShowOneTrial for finding fields of interest and reviewing data structure, respectively.

Author

Ralf Herold ralf.herold@mailbox.org