ctrdata: information on clinical trial registers

Registers of clinical trials from which protocol- and result-related information can be retrieved and analysed with package ctrdata, last updated 2024-12-06.

1 - Overview

EUCTR: The EU Clinical Trials Register contains more than 44,200 clinical trials (at least one investigational medicinal product, IMP; in the European Union and beyond; no new trials, and results of trials continue to be added)

CTIS: The EU Clinical Trials Information System started in January 2023 for new clinical trials. It includes more than 7,900 publicly accessible trials. How to automatically get the CTIS search query URL: here

CTGOV2: ClinicalTrials.gov includes more than 518,000 interventional and observational studies

ISRCTN: The ISRCTN Registry includes more than 25,700 interventional and observational health studies

2 - Notable changes

CTGOV was retired on 2024-06-25; ctrdata subsequently translates CTGOV queries to CTGOV2 queries. The new website (CTGOV2) can be used with ctrdata since 2023-08-27. CTIS was relaunched on 2024-06-17, changing the data structure and search syntax, to which ctrdata was updated. CTIS can be used with ctrdata since 2023-03-25. More information on changes: here.

3 - References

Material	EUCTR	CTGOV2	ISRCTN	CTIS

Home page	link	link	link	link

About	link	link	link	link

Terms & conditions, disclaimer	link	link	link	link

How to search	link	link	link	link

Search interface	link	link	link	link

Expert / advanced search	link	link	link	link

Glossary	link	link	link

FAQ, caveats, issues	link	link	link	link

Definitions of fields	link	link, link	link	link

Example*	link	link	link	link

4 - Example and ctrdata motivation

*The example is an expert search for interventional trials primarily with neonates, investigating infectious conditions. It shows that searches in registers may not be sufficient to identify the sought trials:

The CTGOV2 search retrieves trials conducted exclusively with neonates.

EUCTR retrieves trials with neonates, but not only those exclusively in neonates.

ISRCTN retrieves studies with interventions other than medicines.

CTIS retrieves trials that mention the words neonates and infection. To show search results, see here

To address this, trials can be retrieved with ctrLoadQueryIntoDb into a database collection and in a second step can be selected, based on values of relevant fields of all retrieved trial information, for example:

EUCTR field f115_children_211years for age criteria

ISRCTN field interventions.intervention.interventionType for type of study

CTIS fields ageGroup and authorizedApplication.authorizedPartI.medicalConditions.medicalCondition

ctrdata helps identifying fields with function dbGetFieldsIntoDf.

Author

Ralf Herold ralf.herold@mailbox.org