ctrdata: detailed information on clinical trial registers

Registers of clinical trials that can be accessed with package ctrdata-package as of end 2021

Details

EUCTR: The European Union Clinical Trials Register contains more than 40,000 clinical trials (using one or more medicines as investigational medicinal product, IMP; in Europe and beyond)

CTGOV: ClinicalTrials.gov includes almost 400,000 interventional and observational studies

ISCRTN: The ISRCTN Registry includes more than 21,000 interventional or observational health studies

Material	EUCTR	CTGOV	ISRCTN

Home page	link	link	link

About	link	link	link

Terms and conditions, disclaimer	link	link	link

How to search	link	link	link

Search interface	link	link	link

Glossary	link	link	link

FAQ	link	link	link

Expert / advanced search	link	link	link

Example*	link	link	link

Definitions	link	Protocol, results, names, syntax	link

*The example expert search retrieves interventional trials with neonates, investigating infectious conditions: EUCTR retrieves trials with neonates, but not exclusively. The CTGOV expert search retrieves trials exclusively in neonates. ISRCTN retrieves a small number of studies. Thus, after loading trials with ctrLoadQueryIntoDb into a database collection, corresponding sets of trials need to be defined, based on values of fields of interest (e.g., eligibility.maximum_age from CTGOV and f115_children_211years from EUCTR), which can be obtained from the collection using dbGetFieldsIntoDf.

Author

Ralf Herold ralf.herold@mailbox.org