Find names of fields in the database collection

Given part of the name of a field of interest to the user, this function returns the full field names used in records that were previously loaded into a collection (using ctrLoadQueryIntoDb). The field names can be fed into function dbGetFieldsIntoDf to extract the data from the collection into a data frame.

Usage

dbFindFields(namepart = "", con, verbose = FALSE)

Arguments

namepart

A character string (can include a regular expression, including Perl-style) to be searched among all field names (keys) in the collection, case-insensitive. Use ".*" to find all fields.

con

A connection object, see section `Databases` in ctrdata.

verbose

If TRUE, prints additional information (default FALSE).

Value

Vector of strings with full names of field(s) found, ordered by register and alphabet. Names of the vector elements are the register names for the respective fields.

Only names of fields that have a value in the collection are returned, and the names may be incomplete because they are from sampled records. See here for obtaining a complete set of field names (albeit without register names): https://github.com/rfhb/ctrdata/issues/26#issuecomment-1751452462

Details

In addition to the full names of all child fields (e.g., clinical_results.outcome_list.outcome.measure.class_list.class.title) this function may return names of parent fields (e.g., clinical_results). Data in parent fields is typically complex (nested) and can be converted into individual data elements with dfTrials2Long, and subelements can then be accessed with dfName2Value. For field definitions of the registers, see row "Definition" in ctrdata-registers. Note: Only when dbFindFields is first called after ctrLoadQueryIntoDb, it will take a moment.

Examples

dbc <- nodbi::src_sqlite(
    dbname = system.file("extdata", "demo.sqlite", package = "ctrdata"),
    collection = "my_trials"
)

dbFindFields(namepart = "date", con = dbc)
#> Finding fields in database collection (may take some time)...
#> Field names cached for this session.
#>                                                                                               EUCTR 
#>                                                            "n_date_of_competent_authority_decision" 
#>                                                                                               EUCTR 
#>                                                                "n_date_of_ethics_committee_opinion" 
#>                                                                                               EUCTR 
#>                                                             "p_date_of_the_global_end_of_the_trial" 
#>                                                                                               EUCTR 
#>                                                "trialChanges.globalAmendments.globalAmendment.date" 
#>                                                                                               EUCTR 
#>                                                                "trialInformation.analysisStageDate" 
#>                                                                                               EUCTR 
#>                                                             "trialInformation.globalEndOfTrialDate" 
#>                                                                                               EUCTR 
#>                                                            "trialInformation.primaryCompletionDate" 
#>                                                                                               EUCTR 
#>                                                             "trialInformation.recruitmentStartDate" 
#>                                                                                               EUCTR 
#>                            "x6_date_on_which_this_record_was_first_entered_in_the_eudract_database" 
#>                                                                                               CTGOV 
#>                                                                                   "completion_date" 
#>                                                                                               CTGOV 
#>                                                                                 "completion_date._" 
#>                                                                                               CTGOV 
#>                                                                              "completion_date.type" 
#>                                                                                               CTGOV 
#>                                                                                "last_update_posted" 
#>                                                                                               CTGOV 
#>                                                                              "last_update_posted._" 
#>                                                                                               CTGOV 
#>                                                                           "last_update_posted.type" 
#>                                                                                               CTGOV 
#>                                                                             "last_update_submitted" 
#>                                                                                               CTGOV 
#>                                                                          "last_update_submitted_qc" 
#>                                                                                               CTGOV 
#>                                                                           "primary_completion_date" 
#>                                                                                               CTGOV 
#>                                                                         "primary_completion_date._" 
#>                                                                                               CTGOV 
#>                                                                      "primary_completion_date.type" 
#>                                                                                               CTGOV 
#>                                         "provided_document_section.provided_document.document_date" 
#>                                                                                               CTGOV 
#>                                                                     "required_header.download_date" 
#>                                                                                               CTGOV 
#>                                                                                        "start_date" 
#>                                                                                               CTGOV 
#>                                                                                      "start_date._" 
#>                                                                                               CTGOV 
#>                                                                                   "start_date.type" 
#>                                                                                               CTGOV 
#>                                                                                 "verification_date" 
#>                                                                                              CTGOV2 
#>                                                 "protocolSection.statusModule.completionDateStruct" 
#>                                                                                              CTGOV2 
#>                                            "protocolSection.statusModule.completionDateStruct.date" 
#>                                                                                              CTGOV2 
#>                                            "protocolSection.statusModule.completionDateStruct.type" 
#>                                                                                              CTGOV2 
#>                                             "protocolSection.statusModule.lastUpdatePostDateStruct" 
#>                                                                                              CTGOV2 
#>                                        "protocolSection.statusModule.lastUpdatePostDateStruct.date" 
#>                                                                                              CTGOV2 
#>                                        "protocolSection.statusModule.lastUpdatePostDateStruct.type" 
#>                                                                                              CTGOV2 
#>                                                 "protocolSection.statusModule.lastUpdateSubmitDate" 
#>                                                                                              CTGOV2 
#>                                          "protocolSection.statusModule.primaryCompletionDateStruct" 
#>                                                                                              CTGOV2 
#>                                     "protocolSection.statusModule.primaryCompletionDateStruct.date" 
#>                                                                                              CTGOV2 
#>                                     "protocolSection.statusModule.primaryCompletionDateStruct.type" 
#>                                                                                              CTGOV2 
#>                                           "protocolSection.statusModule.resultsFirstPostDateStruct" 
#>                                                                                              CTGOV2 
#>                                      "protocolSection.statusModule.resultsFirstPostDateStruct.date" 
#>                                                                                              CTGOV2 
#>                                      "protocolSection.statusModule.resultsFirstPostDateStruct.type" 
#>                                                                                              CTGOV2 
#>                                               "protocolSection.statusModule.resultsFirstSubmitDate" 
#>                                                                                              CTGOV2 
#>                                             "protocolSection.statusModule.resultsFirstSubmitQcDate" 
#>                                                                                              CTGOV2 
#>                                                      "protocolSection.statusModule.startDateStruct" 
#>                                                                                              CTGOV2 
#>                                                 "protocolSection.statusModule.startDateStruct.date" 
#>                                                                                              CTGOV2 
#>                                                   "protocolSection.statusModule.statusVerifiedDate" 
#>                                                                                              CTGOV2 
#>                                             "protocolSection.statusModule.studyFirstPostDateStruct" 
#>                                                                                              CTGOV2 
#>                                        "protocolSection.statusModule.studyFirstPostDateStruct.date" 
#>                                                                                              CTGOV2 
#>                                        "protocolSection.statusModule.studyFirstPostDateStruct.type" 
#>                                                                                              CTGOV2 
#>                                                 "protocolSection.statusModule.studyFirstSubmitDate" 
#>                                                                                              CTGOV2 
#>                                               "protocolSection.statusModule.studyFirstSubmitQcDate" 
#>                                                                                              ISRCTN 
#>                                                                        "trialDesign.overallEndDate" 
#>                                                                                              ISRCTN 
#>                                                                      "trialDesign.overallStartDate" 
#>                                                                                                CTIS 
#>                                                    "applications.ctMSCs.activeTrialPeriod.fromDate" 
#>                                                                                                CTIS 
#>                                              "applications.ctMSCs.activeTrialPeriod.trialStartDate" 
#>                                                                                                CTIS 
#>                             "applications.ctMSCs.activeTrialRecruitmentPeriod.recruitmentStartDate" 
#>                                                                                                CTIS 
#>                                                             "applications.ctMSCs.firstDecisionDate" 
#>                                                                                                CTIS 
#>                                                                      "applications.ctMSCs.fromDate" 
#>                                                                                                CTIS 
#>                                       "applications.ctMSCs.organisationInfo.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                                                                        "applications.ctMSCs.toDate" 
#>                                                                                                CTIS 
#>                                                          "applications.ctMSCs.trialPeriod.fromDate" 
#>                                                                                                CTIS 
#>                                                    "applications.ctMSCs.trialPeriod.trialStartDate" 
#>                                                                                                CTIS 
#>                                               "applications.ctMSCs.trialRecruitmentPeriod.fromDate" 
#>                                                                                                CTIS 
#>                                   "applications.ctMSCs.trialRecruitmentPeriod.recruitmentStartDate" 
#>                                                                                                CTIS 
#>                                       "applications.ctMSCsByApplication.activeTrialPeriod.fromDate" 
#>                                                                                                CTIS 
#>                                 "applications.ctMSCsByApplication.activeTrialPeriod.trialStartDate" 
#>                                                                                                CTIS 
#>                "applications.ctMSCsByApplication.activeTrialRecruitmentPeriod.recruitmentStartDate" 
#>                                                                                                CTIS 
#>                                                "applications.ctMSCsByApplication.firstDecisionDate" 
#>                                                                                                CTIS 
#>                                                         "applications.ctMSCsByApplication.fromDate" 
#>                                                                                                CTIS 
#>                          "applications.ctMSCsByApplication.organisationInfo.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                                                           "applications.ctMSCsByApplication.toDate" 
#>                                                                                                CTIS 
#>                                             "applications.ctMSCsByApplication.trialPeriod.fromDate" 
#>                                                                                                CTIS 
#>                                       "applications.ctMSCsByApplication.trialPeriod.trialStartDate" 
#>                                                                                                CTIS 
#>                                  "applications.ctMSCsByApplication.trialRecruitmentPeriod.fromDate" 
#>                                                                                                CTIS 
#>                      "applications.ctMSCsByApplication.trialRecruitmentPeriod.recruitmentStartDate" 
#>                                                                                                CTIS 
#>                                                                         "applications.decisionDate" 
#>                                                                                                CTIS 
#>                                                                      "applications.isDossierUpdate" 
#>                                                                                                CTIS 
#>                                                          "applications.partI.assessmentOutcomeDate" 
#>                                                                                                CTIS 
#>                                                                 "applications.partI.conclusionDate" 
#>                                                                                                CTIS 
#>                                       "applications.partI.productRoleGroupInfos.products.startDate" 
#>                                                                                                CTIS 
#>                                                             "applications.partI.products.startDate" 
#>                                                                                                CTIS 
#>                                      "applications.partI.sponsors.addresses.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                         "applications.partI.sponsors.addresses.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>              "applications.partI.sponsors.contactPointForUnion.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                                                              "applications.partI.sponsors.fromDate" 
#>                                                                                                CTIS 
#>                                   "applications.partI.sponsors.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                    "applications.partI.sponsors.publicContacts.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                "applications.partI.sponsors.scientificContacts.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>               "applications.partI.sponsors.thirdParties.organisationAddress.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>  "applications.partI.sponsors.thirdParties.organisationAddress.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                                                                 "applications.partI.submissionDate" 
#>                                                                                                CTIS 
#>                   "applications.partI.trialDetails.trialInformation.trialDuration.estimatedEndDate" 
#>                                                                                                CTIS 
#>      "applications.partI.trialDetails.trialInformation.trialDuration.estimatedRecruitmentStartDate" 
#>                                                                                                CTIS 
#>                                                     "applications.partIIInfo.applicationStatusDate" 
#>                                                                                                CTIS 
#>                                                            "applications.partIIInfo.conclusionDate" 
#>                                                                                                CTIS 
#>                                        "applications.partIIInfo.mscInfo.activeTrialPeriod.fromDate" 
#>                                                                                                CTIS 
#>                                  "applications.partIIInfo.mscInfo.activeTrialPeriod.trialStartDate" 
#>                                                                                                CTIS 
#>                 "applications.partIIInfo.mscInfo.activeTrialRecruitmentPeriod.recruitmentStartDate" 
#>                                                                                                CTIS 
#>                                                 "applications.partIIInfo.mscInfo.firstDecisionDate" 
#>                                                                                                CTIS 
#>                                                          "applications.partIIInfo.mscInfo.fromDate" 
#>                                                                                                CTIS 
#>                           "applications.partIIInfo.mscInfo.organisationInfo.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                                                            "applications.partIIInfo.mscInfo.toDate" 
#>                                                                                                CTIS 
#>                                              "applications.partIIInfo.mscInfo.trialPeriod.fromDate" 
#>                                                                                                CTIS 
#>                                        "applications.partIIInfo.mscInfo.trialPeriod.trialStartDate" 
#>                                                                                                CTIS 
#>                                   "applications.partIIInfo.mscInfo.trialRecruitmentPeriod.fromDate" 
#>                                                                                                CTIS 
#>                       "applications.partIIInfo.mscInfo.trialRecruitmentPeriod.recruitmentStartDate" 
#>                                                                                                CTIS 
#>                                                            "applications.partIIInfo.submissionDate" 
#>                                                                                                CTIS 
#>                 "applications.partIIInfo.trialSites.organisationAddressInfo.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>    "applications.partIIInfo.trialSites.organisationAddressInfo.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                                      "applications.primarySponsor.addresses.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                         "applications.primarySponsor.addresses.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>              "applications.primarySponsor.contactPointForUnion.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                                                              "applications.primarySponsor.fromDate" 
#>                                                                                                CTIS 
#>                                   "applications.primarySponsor.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                    "applications.primarySponsor.publicContacts.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                "applications.primarySponsor.scientificContacts.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>               "applications.primarySponsor.thirdParties.organisationAddress.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>  "applications.primarySponsor.thirdParties.organisationAddress.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                                                                        "applications.reportingDate" 
#>                                                                                                CTIS 
#>                                                                       "applications.submissionDate" 
#>                                                                                                CTIS 
#>                                                             "applications.validationConclusionDate" 
#>                                                                                                CTIS 
#>                                                                                 "authorizationDate" 
#>                                                                                                CTIS 
#>                                                             "authorizedPartI.assessmentOutcomeDate" 
#>                                                                                                CTIS 
#>                                                                    "authorizedPartI.conclusionDate" 
#>                                                                                                CTIS 
#>                                          "authorizedPartI.productRoleGroupInfos.products.startDate" 
#>                                                                                                CTIS 
#>                                                                "authorizedPartI.products.startDate" 
#>                                                                                                CTIS 
#>                                         "authorizedPartI.sponsors.addresses.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                            "authorizedPartI.sponsors.addresses.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                 "authorizedPartI.sponsors.contactPointForUnion.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                                                                 "authorizedPartI.sponsors.fromDate" 
#>                                                                                                CTIS 
#>                                      "authorizedPartI.sponsors.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                       "authorizedPartI.sponsors.publicContacts.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                   "authorizedPartI.sponsors.scientificContacts.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                  "authorizedPartI.sponsors.thirdParties.organisationAddress.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>     "authorizedPartI.sponsors.thirdParties.organisationAddress.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                                                                    "authorizedPartI.submissionDate" 
#>                                                                                                CTIS 
#>                      "authorizedPartI.trialDetails.trialInformation.trialDuration.estimatedEndDate" 
#>                                                                                                CTIS 
#>         "authorizedPartI.trialDetails.trialInformation.trialDuration.estimatedRecruitmentStartDate" 
#>                                                                                                CTIS 
#>                                                           "authorizedPartsII.applicationStatusDate" 
#>                                                                                                CTIS 
#>                                                                  "authorizedPartsII.conclusionDate" 
#>                                                                                                CTIS 
#>                                              "authorizedPartsII.mscInfo.activeTrialPeriod.fromDate" 
#>                                                                                                CTIS 
#>                                        "authorizedPartsII.mscInfo.activeTrialPeriod.trialStartDate" 
#>                                                                                                CTIS 
#>                       "authorizedPartsII.mscInfo.activeTrialRecruitmentPeriod.recruitmentStartDate" 
#>                                                                                                CTIS 
#>                                                       "authorizedPartsII.mscInfo.firstDecisionDate" 
#>                                                                                                CTIS 
#>                                                                "authorizedPartsII.mscInfo.fromDate" 
#>                                                                                                CTIS 
#>                                 "authorizedPartsII.mscInfo.organisationInfo.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                                                                  "authorizedPartsII.mscInfo.toDate" 
#>                                                                                                CTIS 
#>                                                    "authorizedPartsII.mscInfo.trialPeriod.fromDate" 
#>                                                                                                CTIS 
#>                                              "authorizedPartsII.mscInfo.trialPeriod.trialStartDate" 
#>                                                                                                CTIS 
#>                                         "authorizedPartsII.mscInfo.trialRecruitmentPeriod.fromDate" 
#>                                                                                                CTIS 
#>                             "authorizedPartsII.mscInfo.trialRecruitmentPeriod.recruitmentStartDate" 
#>                                                                                                CTIS 
#>                                                                  "authorizedPartsII.submissionDate" 
#>                                                                                                CTIS 
#>                       "authorizedPartsII.trialSites.organisationAddressInfo.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>          "authorizedPartsII.trialSites.organisationAddressInfo.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                                                       "coSponsors.addresses.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                                          "coSponsors.addresses.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                               "coSponsors.contactPointForUnion.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                                                                               "coSponsors.fromDate" 
#>                                                                                                CTIS 
#>                                                    "coSponsors.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                                     "coSponsors.publicContacts.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                                 "coSponsors.scientificContacts.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                                "coSponsors.thirdParties.organisationAddress.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                   "coSponsors.thirdParties.organisationAddress.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                                                                                      "decisionDate" 
#>                                                                                                CTIS 
#>                                                                                      "eeaStartDate" 
#>                                                                                                CTIS 
#>                                                                                       "lastUpdated" 
#>                                                                                                CTIS 
#>                                                  "memberStatesConcerned.activeTrialPeriod.fromDate" 
#>                                                                                                CTIS 
#>                                            "memberStatesConcerned.activeTrialPeriod.trialStartDate" 
#>                                                                                                CTIS 
#>                           "memberStatesConcerned.activeTrialRecruitmentPeriod.recruitmentStartDate" 
#>                                                                                                CTIS 
#>                                                           "memberStatesConcerned.firstDecisionDate" 
#>                                                                                                CTIS 
#>                                                                    "memberStatesConcerned.fromDate" 
#>                                                                                                CTIS 
#>                                     "memberStatesConcerned.organisationInfo.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                                                                      "memberStatesConcerned.toDate" 
#>                                                                                                CTIS 
#>                                                        "memberStatesConcerned.trialPeriod.fromDate" 
#>                                                                                                CTIS 
#>                                                  "memberStatesConcerned.trialPeriod.trialStartDate" 
#>                                                                                                CTIS 
#>                                             "memberStatesConcerned.trialRecruitmentPeriod.fromDate" 
#>                                                                                                CTIS 
#>                                 "memberStatesConcerned.trialRecruitmentPeriod.recruitmentStartDate" 
#>                                                                                                CTIS 
#>                                       "mscTrialNotificationsInfoList.mscNotificationsListInfo.date" 
#>                                                                                                CTIS 
#>                                 "mscTrialNotificationsInfoList.mscNotificationsListInfo.submitDate" 
#>                                                                                                CTIS 
#>                                                   "primarySponsor.addresses.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                                      "primarySponsor.addresses.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                           "primarySponsor.contactPointForUnion.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                                                                           "primarySponsor.fromDate" 
#>                                                                                                CTIS 
#>                                                "primarySponsor.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                                 "primarySponsor.publicContacts.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                             "primarySponsor.scientificContacts.organisation.isBusinessKeyValidated" 
#>                                                                                                CTIS 
#>                                                           "publicEvaluation.decisions.decisionDate" 
#>                                                                                                CTIS 
#>                                                       "publicEvaluation.partIAssessmentOutcomeDate" 
#>                                                                                                CTIS 
#>              "publicEvaluation.partIIEvaluationList.partIIConclusion.decisionInfoList.decisionDate" 
#>                                                                                                CTIS 
#>                                  "publicEvaluation.partIIEvaluationList.partIIConclusion.eventDate" 
#>                                                                                                CTIS 
#>          "publicEvaluation.partIIEvaluationList.partIIRfiConsiderations.rfiConsiderations.fromDate" 
#>                                                                                                CTIS 
#>  "publicEvaluation.partIIEvaluationList.partIIRfiConsiderations.rfiConsiderations.isDeConsolidated" 
#>                                                                                                CTIS 
#> "publicEvaluation.partIIEvaluationList.partIIRfiConsiderations.rfiConsiderations.rfiSubmissionDate" 
#>                                                                                                CTIS 
#> "publicEvaluation.partIIEvaluationList.partIIRfiConsiderations.rfiConsiderations.sponsorSubmitDate" 
#>                                                                                                CTIS 
#>                                      "publicEvaluation.partIIEvaluationList.partIIRfis.createdDate" 
#>                                                                                                CTIS 
#>                                          "publicEvaluation.partIIEvaluationList.partIIRfis.dueDate" 
#>                                                                                                CTIS 
#>                                     "publicEvaluation.partIIEvaluationList.partIIRfis.responseDate" 
#>                                                                                                CTIS 
#>                                   "publicEvaluation.partIIEvaluationList.partIIRfis.submissionDate" 
#>                                                                                                CTIS 
#>                                "publicEvaluation.partIRfiConsiderations.rfiConsiderations.fromDate" 
#>                                                                                                CTIS 
#>                        "publicEvaluation.partIRfiConsiderations.rfiConsiderations.isDeConsolidated" 
#>                                                                                                CTIS 
#>                       "publicEvaluation.partIRfiConsiderations.rfiConsiderations.rfiSubmissionDate" 
#>                                                                                                CTIS 
#>                       "publicEvaluation.partIRfiConsiderations.rfiConsiderations.sponsorSubmitDate" 
#>                                                                                                CTIS 
#>                                                            "publicEvaluation.partIRfis.createdDate" 
#>                                                                                                CTIS 
#>                                                                "publicEvaluation.partIRfis.dueDate" 
#>                                                                                                CTIS 
#>                                                           "publicEvaluation.partIRfis.responseDate" 
#>                                                                                                CTIS 
#>                                                         "publicEvaluation.partIRfis.submissionDate" 
#>                                                                                                CTIS 
#>                                              "publicEvaluation.validationConclusion.conclusionDate" 
#>                                                                                                CTIS 
#>                           "publicEvaluation.validationRfiConsiderations.rfiConsiderations.fromDate" 
#>                                                                                                CTIS 
#>                   "publicEvaluation.validationRfiConsiderations.rfiConsiderations.isDeConsolidated" 
#>                                                                                                CTIS 
#>                  "publicEvaluation.validationRfiConsiderations.rfiConsiderations.rfiSubmissionDate" 
#>                                                                                                CTIS 
#>                  "publicEvaluation.validationRfiConsiderations.rfiConsiderations.sponsorSubmitDate" 
#>                                                                                                CTIS 
#>                                                       "publicEvaluation.validationRfis.createdDate" 
#>                                                                                                CTIS 
#>                                                           "publicEvaluation.validationRfis.dueDate" 
#>                                                                                                CTIS 
#>                                                      "publicEvaluation.validationRfis.responseDate" 
#>                                                                                                CTIS 
#>                                                    "publicEvaluation.validationRfis.submissionDate" 
#>                                                                                                CTIS 
#>                                                                                       "startDateEU" 
#>                                                                                                CTIS 
#>                                                                                    "submissionDate" 
#>                                                                                                CTIS 
#>                                                                                    "trialStartDate"