Get identifiers of deduplicated trial records

Records for a clinical trial can be loaded from more than one register into a collection. This function returns deduplicated identifiers for all trials in the collection, respecting the register(s) preferred by the user. All registers are recording identifiers also from other registers, which are used by this function to provide a vector of identifiers of deduplicated trials.

Usage

dbFindIdsUniqueTrials(
  preferregister = c("CTGOV2", "EUCTR", "CTGOV", "ISRCTN", "CTIS"),
  prefermemberstate = "BE",
  include3rdcountrytrials = TRUE,
  con,
  verbose = FALSE
)

Arguments

preferregister

A vector of the order of preference for registers from which to generate unique _id's, default c("CTGOV2", "EUCTR", "CTGOV", "ISRCTN", "CTIS")

prefermemberstate

Code of single EU Member State for which records should returned. If not available, a record for BE or lacking this, any random Member State's record for the trial will be returned. For a list of codes of EU Member States, please see vector countriesEUCTR. Specifying "3RD" will return the Third Country record of trials, where available.

include3rdcountrytrials

A logical value if trials should be retained that are conducted exclusively in third countries, that is, outside the European Union. Ignored if prefermemberstate is set to "3RD".

con

A database connection object, created with nodbi. See section `1 - Database connection` in ctrdata.

verbose

If TRUE, prints out the fields of registers used to find corresponding trial records

Value

A named vector with strings of keys (field "_id") of records in the collection that represent unique trials, where names correspond to the register of the record.

Details

Note that the content of records may differ between registers (and, for "EUCTR", between records for different Member States). Such differences are not considered by this function.

Note that the trial concept ".isUniqueTrial" (which uses this function) can be calculated at the time of creating a data frame with dbGetFieldsIntoDf, which often may be the preferred approach.

Examples


dbc <- nodbi::src_sqlite(
  dbname = system.file("extdata", "demo.sqlite", package = "ctrdata"),
  collection = "my_trials",
  flags = RSQLite::SQLITE_RO)

dbFindIdsUniqueTrials(con = dbc)[1:10]
#> Searching for duplicate trials... 
#> - Getting all trial identifiers...
#> , 31 found in collection
#> - Finding duplicates among registers' and sponsor ids...
#> - 2 EUCTR _id were not preferred EU Member State record for 11 trials
#> - Keeping 5 / 3 / 8 / 8 / 5 records from CTGOV2 / EUCTR / CTGOV / ISRCTN / CTIS
#> = Returning keys (_id) of 29 records in collection "my_trials"
#>              ISRCTN              ISRCTN              ISRCTN               EUCTR 
#>          "12949496"          "13281214"          "17473621" "2012-003632-23-SE" 
#>               EUCTR               EUCTR                CTIS                CTIS 
#> "2014-002606-20-PT" "2014-003556-31-SE" "2022-500244-37-00" "2022-501142-30-00" 
#>                CTIS                CTIS 
#> "2023-505613-24-00" "2023-508143-51-01" 

# alternative as of ctrdata version 1.21.0,
# using defaults of dbFindIdsUniqueTrials()
df <- dbGetFieldsIntoDf(
  fields = "keyword",
  calculate = "f.isUniqueTrial",
  con = dbc)
#> To review trial concepts details, call 'help("ctrdata-trial-concepts")'
#> Querying database (2 fields)...
#> Calculating f.isUniqueTrial...                            
#> Searching for duplicate trials... 
#> - Getting all trial identifiers...
#> , 31 found in collection
#> - Finding duplicates among registers' and sponsor ids...
#> - 2 EUCTR _id were not preferred EU Member State record for 11 trials
#> - Keeping 5 / 3 / 8 / 8 / 5 records from CTGOV2 / EUCTR / CTGOV / ISRCTN / CTIS
#> = Returning keys (_id) of 29 records in collection "my_trials"

# using base R
df[df[[".isUniqueTrial"]], ]
#>                  _id
#> 1           12949496
#> 2           13281214
#> 3           17473621
#> 5  2012-003632-23-SE
#> 6  2014-002606-20-PT
#> 8  2014-003556-31-SE
#> 9  2022-500244-37-00
#> 10 2022-501142-30-00
#> 11 2023-505613-24-00
#> 12 2023-508143-51-01
#> 13 2024-510663-34-00
#> 14          20343063
#> 15          61070850
#> 16          76463425
#> 17          80181452
#> 18          88261002
#> 19       NCT00617929
#> 20       NCT01125800
#> 21       NCT01483820
#> 22       NCT01505608
#> 23       NCT01592045
#> 24       NCT02620761
#> 25       NCT03280147
#> 26       NCT03325439
#> 27       NCT03431558
#> 28       NCT03876704
#> 29       NCT04001712
#> 30       NCT04041765
#> 31       NCT05969327
#>                                                                                                   keyword
#> 1                                                                                                    <NA>
#> 2                                                                                                    <NA>
#> 3                                                                                                    <NA>
#> 5                                                                                                    <NA>
#> 6                                                                                                    <NA>
#> 8                                                                                                    <NA>
#> 9                                                                                                    <NA>
#> 10                                                                                                   <NA>
#> 11                                                                                                   <NA>
#> 12                                                                                                   <NA>
#> 13                                                                                                   <NA>
#> 14                                                                                                   <NA>
#> 15                                                                                                   <NA>
#> 16                                                                                                   <NA>
#> 17                                                                                                   <NA>
#> 18                                                                                                   <NA>
#> 19                                                                                                   <NA>
#> 20                                                                                                   <NA>
#> 21                                                                                                   <NA>
#> 22                                                                                                   <NA>
#> 23                                                                                                   <NA>
#> 24                                                                                                   <NA>
#> 25                                                              Neonate / Sepsis / Antibiotics / Duration
#> 26 Electroencephalographic neonatal seizures / Brivaracetam / Epilepsy / ENS / Newborns / Pharmacokinetic
#> 27                                               Bovine Lactoferrin, Neonatal infection, Low Birth Weight
#> 28                                                                                                   <NA>
#> 29                                                                                 early caffeine preterm
#> 30                                                                IgM-enriched Intravenous Immunoglobulin
#> 31                                                                                                   <NA>
#>    .isUniqueTrial
#> 1            TRUE
#> 2            TRUE
#> 3            TRUE
#> 5            TRUE
#> 6            TRUE
#> 8            TRUE
#> 9            TRUE
#> 10           TRUE
#> 11           TRUE
#> 12           TRUE
#> 13           TRUE
#> 14           TRUE
#> 15           TRUE
#> 16           TRUE
#> 17           TRUE
#> 18           TRUE
#> 19           TRUE
#> 20           TRUE
#> 21           TRUE
#> 22           TRUE
#> 23           TRUE
#> 24           TRUE
#> 25           TRUE
#> 26           TRUE
#> 27           TRUE
#> 28           TRUE
#> 29           TRUE
#> 30           TRUE
#> 31           TRUE

if (FALSE) { # \dontrun{
library(dplyr)
df %>% filter(.isUniqueTrial)
} # }