Get identifiers of deduplicated trial records

Records for a clinical trial can be loaded from more than one register into a collection. This function returns deduplicated identifiers for all trials in the collection, respecting the register(s) preferred by the user. All registers are recording identifiers also from other registers, which are used by this function to provide a vector of identifiers of deduplicated trials.

Usage

dbFindIdsUniqueTrials(
  preferregister = c("EUCTR", "CTGOV", "CTGOV2", "ISRCTN", "CTIS"),
  prefermemberstate = "DE",
  include3rdcountrytrials = TRUE,
  con,
  verbose = FALSE
)

Arguments

preferregister

A vector of the order of preference for registers from which to generate unique _id's, default c("EUCTR", "CTGOV", "CTGOV2", "ISRCTN", "CTIS")

prefermemberstate

Code of single EU Member State for which records should returned. If not available, a record for DE or lacking this, any random Member State's record for the trial will be returned. For a list of codes of EU Member States, please see vector countriesEUCTR. Specifying "3RD" will return the Third Country record of trials, where available.

include3rdcountrytrials

A logical value if trials should be retained that are conducted exclusively in third countries, that is, outside the European Union. Ignored if prefermemberstate is set to "3RD".

con

A connection object, see section `Databases` in ctrdata.

verbose

If set to TRUE, prints out which fields of the registers are used to find corresponding trial records

Value

A named vector with strings of keys (field "_id") of records in the collection that represent unique trials, where names correspond to the register of the record.

Details

Note that the content of records may differ between registers (and, for "EUCTR", between records for different Member States). Such differences are not considered by this function.

Examples

dbc <- nodbi::src_sqlite(
    dbname = system.file("extdata", "demo.sqlite", package = "ctrdata"),
    collection = "my_trials"
)

dbFindIdsUniqueTrials(con = dbc)
#> Searching for duplicate trials... 
#>  - Getting all trial identifiers...
#>  (may take some time)...
#> , 36 found in collection
#>  - Finding duplicates among registers' and sponsor ids...
#>  - 10 EUCTR _id were not preferred EU Member State record for 9 trials
#>  - Keeping 2 / 8 / 5 / 8 / 3 records from EUCTR / CTGOV / CTGOV2 / ISRCTN / CTIS
#> = Returning keys (_id) of 26 records in collection "my_trials"
#>              ISRCTN              ISRCTN              ISRCTN               EUCTR 
#>          "12949496"          "13281214"          "17473621" "2012-003632-23-DE" 
#>               EUCTR                CTIS                CTIS                CTIS 
#> "2014-003556-31-SE" "2022-500676-59-00" "2022-501180-40-00" "2022-502516-37-00" 
#>              ISRCTN              ISRCTN              ISRCTN              ISRCTN 
#>          "20343063"          "61070850"          "76463425"          "80181452" 
#>              ISRCTN              CTGOV2              CTGOV2              CTGOV2 
#>          "88261002"       "NCT00617929"       "NCT01125800"       "NCT01483820" 
#>              CTGOV2              CTGOV2               CTGOV               CTGOV 
#>       "NCT01505608"       "NCT01592045"       "NCT02620761"       "NCT03280147" 
#>               CTGOV               CTGOV               CTGOV               CTGOV 
#>       "NCT03325439"       "NCT03431558"       "NCT03876704"       "NCT04001712" 
#>               CTGOV               CTGOV 
#>       "NCT04041765"       "NCT05969327"