Calculate trial concept from various fields and registers

Across registers, calculate a trial concept (e.g., status of recruitment). The concept is defined within ctrdata based on common understanding and any relevant publications. Details of the implementation are printed, if no data frame is specified, to show the peculiarities of the registers for calculating the concept.

Usage

dfCalculateConcept(name = ".*", df = NULL)

Arguments

name

Vector of strings of name of a function(s) to be applied to `df`, or regular expression to list available functions.

df

Optional. Data frame with fields needed to apply function `name`.

Value

Data frame with additional column(s) `name` as calculated applying the function(s) `name`. If `df` is `NULL` (default), either a list of functions corresponding to `name` or, if exactly one function is identified, prints details of the function `name` to explain the implementation of the concept and to return invisibly the names of fields needed for its calculation.

Details

The names of the functions can also be used for the argument `calculate` in dbGetFieldsIntoDf to achieve the same purpose already at the time that a data frame is generated from the trial collection. This approach will in most cases be more efficient and preferred over using dfCalculateConcept.

Note that the functions do not operate on historic versions (CTGOV2) but only on information from the latest record.

Examples

# list available functions to calculate trial concepts
dfCalculateConcept()
#>  [1] "f.controlType"                "f.isMedIntervTrial"          
#>  [3] "f.isPlatformTrial"            "f.isUniqueTrial"             
#>  [5] "f.numSites"                   "f.numTestArmsSubstances"     
#>  [7] "f.primaryEndpointDescription" "f.primaryEndpointResults"    
#>  [9] "f.resultsDate"                "f.sampleSize"                
#> [11] "f.sponsorType"                "f.startDate"                 
#> [13] "f.statusRecruitment"          "f.trialObjectives"           
#> [15] "f.trialPhase"                 "f.trialPopulation"           

# list names of functions for regular expression
dfCalculateConcept(name = "date")
#> [1] "f.resultsDate" "f.startDate"  

# use package trial database collection
dbc <- nodbi::src_sqlite(
  dbname = system.file("extdata", "demo.sqlite", package = "ctrdata"),
  collection = "my_trials", flags = RSQLite::SQLITE_RO)
#> RSQLite version has enabled accelerating docdb_create() and docdb_update() functions when used with value = <NDJSON file name>.

# calculate trial concepts when creating data frame
trialsDf <- dbGetFieldsIntoDf(
  calculate = c("f.statusRecruitment", "f.controlType"),
  con = dbc)
#> Querying database (26 fields)...
#> Calculating f.statusRecruitment...                            
#> Calculating f.controlType...                            

# or calculate trial concepts for existing data frame
trialsDf <- dbGetFieldsIntoDf(
  fields = unlist(f.statusRecruitment()),
  con = dbc)
#> Querying database (11 fields)...
dfCalculateConcept(
  name = "f.statusRecruitment",
  df = trialsDf)
#> Calculating f.statusRecruitment...                            
#>                  _id trialInformation.globalEndOfTrialPremature
#> 1           12949496                                         NA
#> 2           13281214                                         NA
#> 3           17473621                                         NA
#> 4  2012-003632-23-CZ                                      FALSE
#> 5  2012-003632-23-SE                                      FALSE
#> 6  2014-002606-20-PT                                         NA
#> 7  2014-003556-31-GB                                      FALSE
#> 8  2014-003556-31-SE                                      FALSE
#> 9  2022-501142-30-00                                         NA
#> 10 2023-505613-24-00                                         NA
#> 11 2024-510663-34-00                                         NA
#> 12          20343063                                         NA
#> 13          61070850                                         NA
#> 14          76463425                                         NA
#> 15          80181452                                         NA
#> 16          88261002                                         NA
#> 17       NCT00617929                                         NA
#> 18       NCT01125800                                         NA
#> 19       NCT01483820                                         NA
#> 20       NCT01505608                                         NA
#> 21       NCT01592045                                         NA
#> 22       NCT02620761                                         NA
#> 23       NCT03280147                                         NA
#> 24       NCT03325439                                         NA
#> 25       NCT03431558                                         NA
#> 26       NCT03876704                                         NA
#> 27       NCT04001712                                         NA
#> 28       NCT04041765                                         NA
#> 29       NCT05969327                                         NA
#>    trialInformation.isGlobalEndOfTrialReached p_end_of_trial_status
#> 1                                          NA                  <NA>
#> 2                                          NA                  <NA>
#> 3                                          NA                  <NA>
#> 4                                        TRUE             Completed
#> 5                                        TRUE             Completed
#> 6                                          NA               Ongoing
#> 7                                        TRUE                  <NA>
#> 8                                        TRUE             Completed
#> 9                                          NA                  <NA>
#> 10                                         NA                  <NA>
#> 11                                         NA                  <NA>
#> 12                                         NA                  <NA>
#> 13                                         NA                  <NA>
#> 14                                         NA                  <NA>
#> 15                                         NA                  <NA>
#> 16                                         NA                  <NA>
#> 17                                         NA                  <NA>
#> 18                                         NA                  <NA>
#> 19                                         NA                  <NA>
#> 20                                         NA                  <NA>
#> 21                                         NA                  <NA>
#> 22                                         NA                  <NA>
#> 23                                         NA                  <NA>
#> 24                                         NA                  <NA>
#> 25                                         NA                  <NA>
#> 26                                         NA                  <NA>
#> 27                                         NA                  <NA>
#> 28                                         NA                  <NA>
#> 29                                         NA                  <NA>
#>     last_known_status overall_status protocolSection.statusModule.overallStatus
#> 1                <NA>           <NA>                                       <NA>
#> 2                <NA>           <NA>                                       <NA>
#> 3                <NA>           <NA>                                       <NA>
#> 4                <NA>           <NA>                                       <NA>
#> 5                <NA>           <NA>                                       <NA>
#> 6                <NA>           <NA>                                       <NA>
#> 7                <NA>           <NA>                                       <NA>
#> 8                <NA>           <NA>                                       <NA>
#> 9                <NA>           <NA>                                       <NA>
#> 10               <NA>           <NA>                                       <NA>
#> 11               <NA>           <NA>                                       <NA>
#> 12               <NA>           <NA>                                       <NA>
#> 13               <NA>           <NA>                                       <NA>
#> 14               <NA>           <NA>                                       <NA>
#> 15               <NA>           <NA>                                       <NA>
#> 16               <NA>           <NA>                                       <NA>
#> 17               <NA>           <NA>                                 TERMINATED
#> 18               <NA>           <NA>                                  COMPLETED
#> 19               <NA>           <NA>                                 TERMINATED
#> 20               <NA>           <NA>                                 TERMINATED
#> 21               <NA>           <NA>                                  COMPLETED
#> 22               <NA>     Terminated                                       <NA>
#> 23               <NA>      Completed                                       <NA>
#> 24               <NA>     Terminated                                       <NA>
#> 25         Recruiting Unknown status                                       <NA>
#> 26         Recruiting Unknown status                                       <NA>
#> 27               <NA>      Completed                                       <NA>
#> 28 Not yet recruiting Unknown status                                       <NA>
#> 29               <NA>      Completed                                       <NA>
#>    participants.recruitmentEnd participants.recruitmentStart
#> 1                   2020-01-01                    2015-11-01
#> 2                   2020-01-01                    2016-09-01
#> 3                   2024-02-01                    2023-05-01
#> 4                         <NA>                          <NA>
#> 5                         <NA>                          <NA>
#> 6                         <NA>                          <NA>
#> 7                         <NA>                          <NA>
#> 8                         <NA>                          <NA>
#> 9                         <NA>                          <NA>
#> 10                        <NA>                          <NA>
#> 11                        <NA>                          <NA>
#> 12                  2013-12-31                    2011-01-01
#> 13                  2011-12-31                    2007-01-10
#> 14                  2015-05-31                    2013-06-01
#> 15                  2012-03-31                    2010-04-01
#> 16                  2017-09-28                    2013-09-01
#> 17                        <NA>                          <NA>
#> 18                        <NA>                          <NA>
#> 19                        <NA>                          <NA>
#> 20                        <NA>                          <NA>
#> 21                        <NA>                          <NA>
#> 22                        <NA>                          <NA>
#> 23                        <NA>                          <NA>
#> 24                        <NA>                          <NA>
#> 25                        <NA>                          <NA>
#> 26                        <NA>                          <NA>
#> 27                        <NA>                          <NA>
#> 28                        <NA>                          <NA>
#> 29                        <NA>                          <NA>
#>    participants.recruitmentStatusOverride ctPublicStatusCode   ctStatus
#> 1                                 Stopped                 NA       <NA>
#> 2                                    <NA>                 NA       <NA>
#> 3                                    <NA>                 NA       <NA>
#> 4                                    <NA>                 NA       <NA>
#> 5                                    <NA>                 NA       <NA>
#> 6                                    <NA>                 NA       <NA>
#> 7                                    <NA>                 NA       <NA>
#> 8                                    <NA>                 NA       <NA>
#> 9                                    <NA>                  4 Authorised
#> 10                                   <NA>                  4 Authorised
#> 11                                   <NA>                  4 Authorised
#> 12                                   <NA>                 NA       <NA>
#> 13                                   <NA>                 NA       <NA>
#> 14                                   <NA>                 NA       <NA>
#> 15                                   <NA>                 NA       <NA>
#> 16                                   <NA>                 NA       <NA>
#> 17                                   <NA>                 NA       <NA>
#> 18                                   <NA>                 NA       <NA>
#> 19                                   <NA>                 NA       <NA>
#> 20                                   <NA>                 NA       <NA>
#> 21                                   <NA>                 NA       <NA>
#> 22                                   <NA>                 NA       <NA>
#> 23                                   <NA>                 NA       <NA>
#> 24                                   <NA>                 NA       <NA>
#> 25                                   <NA>                 NA       <NA>
#> 26                                   <NA>                 NA       <NA>
#> 27                                   <NA>                 NA       <NA>
#> 28                                   <NA>                 NA       <NA>
#> 29                                   <NA>                 NA       <NA>
#>    .statusRecruitment
#> 1               other
#> 2           completed
#> 3           completed
#> 4           completed
#> 5           completed
#> 6             ongoing
#> 7           completed
#> 8           completed
#> 9             ongoing
#> 10            ongoing
#> 11            ongoing
#> 12          completed
#> 13          completed
#> 14          completed
#> 15          completed
#> 16          completed
#> 17              other
#> 18          completed
#> 19              other
#> 20              other
#> 21          completed
#> 22              other
#> 23          completed
#> 24              other
#> 25            ongoing
#> 26            ongoing
#> 27          completed
#> 28            ongoing
#> 29          completed