Get value for variable of interest

Get information for variable of interest (e.g., clinical endpoints) from long data frame of protocol- or result-related trial information as returned by dfTrials2Long. Parameters `valuename`, `wherename` and `wherevalue` are matched using Perl regular expressions and ignoring case.

Usage

dfName2Value(df, valuename = "", wherename = "", wherevalue = "")

Arguments

df

A data frame (or tibble) with four columns (`_id`, `identifier`, `name`, `value`) as returned by dfTrials2Long

valuename

A character string for the name of the field that holds the value of the variable of interest (e.g., a summary measure such as "endPoints.*tendencyValue.value")

wherename

(optional) A character string to identify the variable of interest among those that repeatedly occur in a trial record (e.g., "endPoints.endPoint.title")

wherevalue

(optional) A character string with the value of the variable identified by `wherename` (e.g., "response")

Value

A data frame (or tibble, if tibble is loaded) that includes the values of interest, with columns `_id`, `identifier`, `name`, `value` (and `where`, with the contents of `wherevalue` found at `wherename`). Contents of `value` are strings unless all its elements are numbers. The `identifier` is generated by function

dfTrials2Long to identify matching elements, e.g. endpoint descriptions and measurements.

Examples

dbc <- nodbi::src_sqlite(
    dbname = system.file("extdata", "demo.sqlite", package = "ctrdata"),
    collection = "my_trials"
)

dfwide <- dbGetFieldsIntoDf(
    fields = c(
        ## ctgov - typical results fields
        # "clinical_results.baseline.analyzed_list.analyzed.count_list.count",
        # "clinical_results.baseline.group_list.group",
        # "clinical_results.baseline.analyzed_list.analyzed.units",
        "clinical_results.outcome_list.outcome",
        "study_design_info.allocation",
        ## euctr - typical results fields
        # "trialInformation.fullTitle",
        # "baselineCharacteristics.baselineReportingGroups.baselineReportingGroup",
        # "trialChanges.hasGlobalInterruptions",
        # "subjectAnalysisSets",
        # "adverseEvents.seriousAdverseEvents.seriousAdverseEvent",
        "endPoints.endPoint",
        "subjectDisposition.recruitmentDetails"
    ), con = dbc
)
#> clinical_results.outcome_list.outcome... 
#>     
#> . 
#> . 
#> 
#> study_design_info.allocation... 
#>     
#> . 
#> 
#> endPoints.endPoint... 
#>     
#> . 
#> 
#> subjectDisposition.recruitmentDetails... 
#>     
#> . 

dflong <- dfTrials2Long(df = dfwide)
#> clinical_results.outcome_list.outcome                                                                                                                                                                   
#> study_design_info.allocation                                                                                                                                                                            
#> endPoints.endPoint                                                                                                                                                                                      
#> subjectDisposition.recruitmentDetails                                                                                                                                                                   
#>                                                                                                                                                                                                         
#> . 
#> . 
#> . 
#> . 
#> . 
#> . 
#> . 
#> . 
#> 
#> Total 21146 rows, 90 unique names of variables

## get values for the endpoint 'response'
dfName2Value(
    df = dflong,
    valuename = paste0(
        "clinical_results.*measurement.value|",
        "clinical_results.*outcome.measure.units|",
        "endPoints.endPoint.*tendencyValue.value|",
        "endPoints.endPoint.unit"
    ),
    wherename = paste0(
        "clinical_results.*outcome.measure.title|",
        "endPoints.endPoint.title"
    ),
    wherevalue = "response"
)
#> Returning values for 7 out of 46 trials
#> # A tibble: 77 × 5
#>    `_id`             identifier name                                                                 value where
#>    <chr>             <chr>      <chr>                                                                <chr> <chr>
#>  1 2012-003632-23-AT 1          endPoints.endPoint.unit                                              Days  Time…
#>  2 2012-003632-23-AT 1          endPoints.endPoint.subjectAnalysisSetReportingGroups.subjectAnalysi… 7.0   Time…
#>  3 2012-003632-23-AT 2          endPoints.endPoint.unit                                              At l… Dura…
#>  4 2012-003632-23-AT 6          endPoints.endPoint.unit                                              Over… Over…
#>  5 2012-003632-23-AT 6.1        endPoints.endPoint.subjectAnalysisSetReportingGroups.subjectAnalysi… 0.63  Over…
#>  6 2012-003632-23-AT 6.2        endPoints.endPoint.subjectAnalysisSetReportingGroups.subjectAnalysi… 0     Over…
#>  7 2012-003632-23-AT 6.3        endPoints.endPoint.subjectAnalysisSetReportingGroups.subjectAnalysi… 0.65  Over…
#>  8 2012-003632-23-AT 6.4        endPoints.endPoint.subjectAnalysisSetReportingGroups.subjectAnalysi… 0.59  Over…
#>  9 2012-003632-23-AT 6.5        endPoints.endPoint.subjectAnalysisSetReportingGroups.subjectAnalysi… 0.60  Over…
#> 10 2012-003632-23-AT 8          endPoints.endPoint.unit                                              Perc… Cumu…
#> # ℹ 67 more rows